### 2024-07-15 #6990 @editorialbot
[PRE REVIEW]: PhaseFieldX: An Open-Source Framework for Advanced Phase-Field Simulations
**Submitting author:** @CastillonMiguel (<a href="http://orcid.org/0009-0008-6383-7103">Miguel Castillón de Miguel</a>)
**Repository:** https://github.com/CastillonMiguel/phasefieldx
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.2
