### 2024-07-15 1cfe851
Remove 'https://doi.org/' prefix from DOI in bibliography entries
