+++ joss/paper.bib
@@ -48,7 +48,7 @@ number = {45},
 pages = {2765-2778},
 year = {2010},
 issn = {0045-7825},
-doi = {https://doi.org/10.1016/j.cma.2010.04.011},
+doi = {10.1016/j.cma.2010.04.011},
 url = {https://www.sciencedirect.com/science/article/pii/S0045782510001283},
 author    = {Miehe, C. and Hofacker, M. and Welschinger, F.},
 }
@@ -61,7 +61,7 @@ number = {8},
 pages = {1209-1229},
 year = {2009},
 issn = {0022-5096},
-doi = {https://doi.org/10.1016/j.jmps.2009.04.011},
+doi = {10.1016/j.jmps.2009.04.011},
 url = {https://www.sciencedirect.com/science/article/pii/S0022509609000659},
 author    = {Amor, H. and Marigo, J.-J. and Maurini, C.},
 keywords = {Fracture, Energy methods, Free-discontinuity problems, Contact mechanics, Finite elements}
@@ -90,7 +90,7 @@ volume = {361},
 pages = {112731},
 year = {2020},
 issn = {0045-7825},
-doi = {https://doi.org/10.1016/j.cma.2019.112731},
+doi = {10.1016/j.cma.2019.112731},
 url = {https://www.sciencedirect.com/science/article/pii/S0045782519306218},
 author    = {Carrara, P. and Ambati, M. and Alessi, R. and De Lorenzis, L.},
 keywords = {Brittle fracture, Fatigue, Modeling, Paris law, Phase-field models, Wöhler curve},
