### 2024-07-31 01ec8c2
Add .geo file for structured rectangular mesh and corresponding Python visualization script
