### 2024-07-31 5d0fba7
Update Python script for mesh visualization
