+++ README.rst
@@ -8,6 +8,9 @@ PhaseFieldX
 
 Welcome to **PhaseFieldX**. `documentation <https://phasefieldx.readthedocs.io/en/latest/index.html>`_
 
+.. image:: https://readthedocs.org/projects/phasefieldx/badge/?version=latest
+    :target: https://phasefieldx.readthedocs.io/en/latest/?badge=latest
+    :alt: ∑
 
 .. image:: https://img.shields.io/pypi/v/phasefieldx
     :target: https://pypi.org/project/phasefieldx/
