### 2024-08-21 c29b9c5
added example: Structured Mesh with Gmsh API
