### 2024-08-21 67f3985
Add Contributor Covenant Code of Conduct
