+++ docs/requirements.txt
@@ -8,7 +8,7 @@ jupyterlab
 lxml
 meshio
 mpi4py==3.1.6
-mypy
+mypy==1.11.0
 mypy-extensions
 numpydoc
 osmnx
