+++ docs/requirements.txt
@@ -7,8 +7,8 @@ jupyter_sphinx
 jupyterlab
 lxml
 meshio
-mpi4py==3.1.6
-mypy==1.11.0
+mpi4py
+mypy
 mypy-extensions
 numpydoc
 osmnx
