### 2024-08-26 61cad2b
Add example: Center Notched Tension Test
