+++ examples/PhaseFieldFracture/plot_1715.py
@@ -1,0 +1,264 @@
+"""
+.. _ref_1715:
+
+Symmetry: Center notched tension test
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+The model represents a square plate with a central crack, as shown in the figure below. The bottom part is fixed in all directions, while the upper part can slide vertically. A vertical displacement is applied at the top. The geometry and boundary conditions are depicted in the figure. We discretize the model with quadrilateral elements.
+
+.. note::
+   In this case, only one quarter of the model will be considered due to symmetry. Additionally, a regular mesh will be used.
+
+.. code-block::
+
+   #           u/\/\/\/\/\/\       u/\/\/\/\/\/\  
+   #            ||||||||||||        ||||||||||||
+   #            *----------*    o|\ *----------*  
+   #            |          |    o|/ |          |
+   #            | 2a=1.0   |    o|\ | a=0.5    |
+   #            |   ----   |    o|/ *----------* 
+   #            |          |             /_\/_\ 
+   #            |          |            oo oo oo 
+   #            *----------*
+   #            /_\/_\/_\/_\       
+   #     |Y    /////////////
+   #     |                
+   #     *---X           
+
+   
++----+---------+--------+
+|    | VALUE   | UNITS  |
++====+=========+========+
+| E  | 210     | kN/mm2 |
++----+---------+--------+
+| nu | 0.3     | [-]    |
++----+---------+--------+
+| Gc | 0.0027  | kN/mm2 |
++----+---------+--------+
+| l  | 0.015   | mm     |
++----+---------+--------+
+
+
+"""
+
+###############################################################################
+# Import necessary libraries
+# --------------------------
+import numpy as np
+import matplotlib.pyplot as plt
+import pyvista as pv
+import dolfinx
+import mpi4py 
+import petsc4py
+import os
+
+
+###############################################################################
+# Import from phasefieldx package
+# -------------------------------
+from phasefieldx.Element.Phase_Field_Fracture.Input import Input
+from phasefieldx.Element.Phase_Field_Fracture.solver.solver import solve
+from phasefieldx.Boundary.boundary_conditions import bc_x, bc_y, get_ds_bound_from_marker
+from phasefieldx.PostProcessing.ReferenceResult import AllResults
+
+
+###############################################################################
+# Parameters Definition
+# ---------------------
+# Here, we define the class containing the input parameters. The material parameters are:
+# Young's modulus $E = 210$, $\text{kN/mm}^2$, Poisson's ratio $\nu = 0.3$, 
+# critical energy release rate $G_c = 0.0027$, $\text{kN/mm}^2$, and length scale parameter $l = 0.025$, $\text{mm}$. 
+# We consider isotropic degradation where all the energy is degraded and irreversibility as proposed by Miehe.
+Data = Input(E=210.0,   # young modulus
+             nu=0.3,    # poisson
+             Gc=0.0027, # critical energy release rate
+             l=0.025,   # lenght scale parameter
+             degradation="isotropic", # "isotropic" "anisotropic"
+             split_energy="no",       # "spectral" "deviatoric"
+             degradation_function="quadratic",
+             irreversibility="miehe", # "miehe"
+             fatigue=False,
+             fatigue_degradation_function="asymptotic",
+             fatigue_val=0.05625,
+             k=0.0,
+             min_stagger_iter=2,
+             max_stagger_iter=600,
+             stagger_error_tol=1e-8,
+             save_solution_xdmf=False,
[…]
