### 2024-08-26 14d39f2
Add JOSS status badge to README
