### 2024-09-04 b1c37ec
Correct link syntax in contribution section from Markdown to reStructuredText
