### 2024-09-09 eb16892
add "pv.start_xvfb()" for online docs
