+++ examples/PhaseFieldFracture/plot_1714.py
@@ -211,6 +211,7 @@ S.set_color('b')
 # The phase-field result saved in the .vtu file is shown. 
 # For this, the file is loaded using PyVista.
 file_vtu = pv.read(os.path.join(Data.results_folder_name, "paraview-solutions_vtu", "phasefieldx_p0_000065.vtu"))
+pv.start_xvfb()
 file_vtu.plot(scalars='phi', cpos='xy', show_scalar_bar=True, show_edges=False)
 
 
@@ -220,6 +221,7 @@ file_vtu.plot(scalars='phi', cpos='xy', show_scalar_bar=True, show_edges=False)
 # The displacements results saved in the .vtu file are shown. 
 # For this, the file is loaded using PyVista.
 file_vtu = pv.read(os.path.join(Data.results_folder_name, "paraview-solutions_vtu", "phasefieldx_p0_000065.vtu"))
+pv.start_xvfb()
 file_vtu.plot(scalars='u', cpos='xy', show_scalar_bar=True, show_edges=False)
 
 plt.show()
