### 2024-09-27 #6990 @zhubonan
Hello  @IgorBaratta Would you be able to review this submission? The process of reviewing for JOSS is unique and takes place on GitHub. This involves reviewing the manuscript and installing/testing/reading the underlying code. You can learn more about it here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#reviewing-for-joss

Please let me know if you are able/unable to review. Your response is greatly appreciated. Thanks in advance.
