### 2024-10-01 #7307 @editorialbot
[REVIEW]: PhaseFieldX: An Open-Source Framework for Advanced Phase-Field Simulations
**Submitting author:** @CastillonMiguel (<a href="http://orcid.org/0009-0008-6383-7103">Miguel Castillón de Miguel</a>)
**Repository:** https://github.com/CastillonMiguel/phasefieldx
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
