### 2024-10-01 #7307 @zhubonan
Hello @finsberg & @IgorBaratta 👏 , thanks for agreeing to review this submission. The review process is checklist based. Each of you will have a separate checklist that you should update when carrying out your review.

To get started, please run this command in a separate comment to create the checklist:

The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. If you have any question, please let me know.

If you want to make comments and suggestions in the review process, you can open issues in the project repo and mention this review issue to keep things in track.

Please finish the initial review in two weeks time, if possible, but do let me know if you think more time is needed. Thanks in advance!
