### 2024-10-06 CastillonMiguel/phasefieldx#1 @finsberg
[JOSS REVIEW]: Add Contributing guidelines
As a part the JOSS submission you should provide clear guidelines for third-parties wishing to:

- Contribute to the software
- Report issues or problems with the software
- Seek support

While you have mentioned in the README that contributors should engage through the GitHub repository, I do think it would be good with more information about this. For example you could add a [`CONTRIBUTING.md`](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) with instructions on whether authors should first open an issue, do you want contributors to add tests / documentation, should contributors follow any particular code styles etc (you can of course also provide this information elsewhere if you feel that is more natural).

For users that want to report issues or seek support, it might be good to add a link to the issue tracker.

This issue is related to the review here: https://github.com/openjournals/joss-reviews/issues/7307
