### 2024-10-07 CastillonMiguel/phasefieldx#2 @finsberg
[JOSS REVIEW]: Add `phasefieldx` to conda-forge
To simplify the installation of `phasefieldx` you should add it to conda-forge, see https://conda-forge.org/docs/maintainer/adding_pkgs/ on how to add a package to conda-forge.

You can checkout e.g [adios4dolfinx](https://github.com/conda-forge/adios4dolfinx-feedstock/blob/main/recipe/meta.yaml) as an example of a pure python package on conda-forge that depends on dolfinx.

This issue is related to the review here: https://github.com/openjournals/joss-reviews/issues/7307
