+++ README.rst
@@ -105,6 +105,6 @@ For detailed API documentation, including class references, function definitions
 
 Contributions and Feedback
 --------------------------
-We welcome contributions and feedback from the community to enhance the code's functionality, reliability, and user experience. Engage with us through our `GitHub repository <https://github.com/CastillonMiguel/phasefieldx>`_ to share your insights.
+We welcome contributions and feedback from the community to enhance the code's functionality, reliability, and user experience. Engage with us through our `GitHub repository <https://phasefieldx.readthedocs.io/en/latest/extras/DeveloperNotes/main.html>`_ to share your insights.
 
 Thank you for choosing our Phase-Field Fracture simulation code. We trust this tool will prove invaluable in advancing your understanding of fracture mechanics and its practical applications.
