+++ docs/source/extras/index.rst
@@ -1,6 +1,9 @@
 Extras
 ******
 
+For more information on contributing and how you can support the development of **PhaseFieldX**, refer to the :ref:`developer_notes` section.
+
+
 .. toctree::
    :hidden:
 
