+++ docs/source/extras/DeveloperNotes/main.rst
@@ -1,0 +1,3 @@
+.. _developer_notes:
+
+.. include:: ../../../../CONTRIBUTING.rst
