### 2024-10-07 CastillonMiguel/phasefieldx#1 @CastillonMiguel
I appreciate your suggestions on providing clearer guidelines for contributors and users.

I have created a [CONTRIBUTING.rst](https://github.com/CastillonMiguel/phasefieldx/blob/main/CONTRIBUTING.rst) file that includes detailed instructions on how to contribute, report issues, and seek support, along with information on code style and testing.

Additionally, I’ve added a link to the issue tracker in the README for easier access.

See commit [Add Contributing guidelines](https://github.com/CastillonMiguel/phasefieldx/commit/163d73417b956ab28feaf02fc365739c5f70d35a), and [Update README](https://github.com/CastillonMiguel/phasefieldx/commit/4b15392fd4ad7c56000d695ab31ff6eaf3d83d87) for the changes.
