### 2024-10-07 CastillonMiguel/phasefieldx#3 @finsberg
[JOSS REVIEW]: Add automated tests
The documentation for the library looks really nice and it seems to me that all the examples are actually run when you build the documentation on readthedocs. In this way you could argue that the examples in the documentation actually serves as tests.
However, I think it would be good to have some automated tests that can run whenever someone pushes to the repo or open a pull request so that potential contributors (and yourself) can be notified if you break some functionality. I am not so familiar with readthedocs, and it might be that you will get notified if building the docs fails?

Since you already have the examples in place, you could just create a github workflow that runs the examples in CI. This way you can also more easily upgrade to a new version by just trying a new docker image.

If the examples also run in parallel, then if would also be good to add a test for that.

I did something similar here https://github.com/RangamaniLabUCSD/smart/blob/development/.github/workflows/run_examples_mpi.yml (allthough this is using Legacy FEniCS).

This issue is related to the review here: https://github.com/openjournals/joss-reviews/issues/7307
