### 2024-10-07 CastillonMiguel/phasefieldx#1 @finsberg
Thanks @CastillonMiguel! This looks very good. One question related to the guide is this line: _Add tests for any new features or bug fixes_. To me it is not quite clear how a user should add a test.
