### 2024-10-07 CastillonMiguel/phasefieldx#4 @finsberg
[JOSS REVIEW]: SyntaxError: f-string: unmatched '('
I am trying to run the examples in the example folder, e.g in the `AllenCahn` folder I get the following error
```
root@962b66de34de:/home/shared/examples/AllenCahn# python3 plot_4000.py
Traceback (most recent call last):
  File "/home/shared/examples/AllenCahn/plot_4000.py", line 67, in <module>
    from phasefieldx.Element.Allen_Cahn.Input import Input
  File "/home/shared/src/phasefieldx/__init__.py", line 12, in <module>
    from .Element import *
  File "/home/shared/src/phasefieldx/Element/__init__.py", line 4, in <module>
    from .Allen_Cahn import *
  File "/home/shared/src/phasefieldx/Element/Allen_Cahn/__init__.py", line 6, in <module>
    from .solver import *
  File "/home/shared/src/phasefieldx/Element/Allen_Cahn/solver/__init__.py", line 3, in <module>
    from .solver import *
  File "/home/shared/src/phasefieldx/Element/Allen_Cahn/solver/solver.py", line 16, in <module>
    from phasefieldx.Logger.library_versions import set_logger, log_library_versions, log_system_info, log_end_analysis, log_model_information
  File "/home/shared/src/phasefieldx/Logger/__init__.py", line 3, in <module>
    from .library_versions import *
  File "/home/shared/src/phasefieldx/Logger/library_versions.py", line 117
    logger.info(f"Analysis finished on {time.strftime("%a %b %d %H:%M:%S %Y", time.localtime())}")
                                                       ^
SyntaxError: f-string: unmatched '('
```

Here I am using python3.10.12 inside docker using the docker image `ghcr.io/fenics/dolfinx/dolfinx:v0.8.0`.
I belive this might be related to [PEP701](https://peps.python.org/pep-0701/), which is implemented for python version 3.12. In your `pyproject.toml` it sais that `requires-python = ">=3.8"`.

By the way, there is an easy fix to this problem by just replacing the inner most quotes from `"` to.`'`.

Thie is related to the joss review here: https://github.com/openjournals/joss-reviews/issues/7307
