### 2024-10-08 CastillonMiguel/phasefieldx#3 @CastillonMiguel
Yes, I get notified if the documentation build fails. When I push changes, ReadTheDocs builds all the documentation and runs all the examples. If something is wrong, the documentation build fails, and I receive an error.  However, most of the current examples have a high computational cost, so I preload the results directly, meaning they don’t serve as tests.

The best approach might be to add some new, simple automated tests to help catch potential errors. I will implement these new tests. Thanks!
