### 2024-10-08 CastillonMiguel/phasefieldx#1 @CastillonMiguel
For now, the way to test is by running and checking the new examples. I haven’t added detailed instructions yet. Regarding automated tests, I think the best approach is to first complete  [issue #3](https://github.com/CastillonMiguel/phasefieldx/issues/3), and then I will update the guide with the correct instructions for adding new tests.
