### 2024-10-08 a4d7dbc
Fix: Updated pyproject.toml to use single quotes for requires-python field (PEP701)
