+++ pyproject.toml
@@ -10,7 +10,7 @@ authors = [
 ]
 description = "PhaseFieldX: An Open-Source Framework for Advanced Phase-Field Simulations"
 readme = "README.rst"
-requires-python = ">=3.8"
+requires-python = '>=3.8'
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
