### 2024-10-08 CastillonMiguel/phasefieldx#2 @CastillonMiguel
For now, it’s possible to install phasefieldx via pip using pip install phasefieldx. However, I understand the benefit of adding it to conda-forge for easier installation and wider compatibility with conda environments. I’ll look into adding phasefieldx to the conda-forge channel following the guidelines you provided. Thanks for pointing me to adios4dolfinx as an example; it will be helpful as a reference.
