### 2024-10-08 CastillonMiguel/phasefieldx#4 @CastillonMiguel
Thank you for pointing this out! I’ve updated the pyproject.toml file to use single quotes for the requires-python field, changing it from ">=3.8" to '>=3.8'   https://github.com/CastillonMiguel/phasefieldx/commit/a4d7dbc3083123ca13a10f772d38721a8c916c00
