### 2024-10-08 CastillonMiguel/phasefieldx#4 @finsberg
Hi @CastillonMiguel. Sorry for the confusion. I meant chaning
```python
logger.info(f"Analysis finished on {time.strftime("%a %b %d %H:%M:%S %Y", time.localtime())}")
```
to
```python
logger.info(f"Analysis finished on {time.strftime('%a %b %d %H:%M:%S %Y', time.localtime())}")
```
The former is only possible in python3.12 and above due to PEP701
