### 2024-10-09 bf7b879
Add unit tests for Allen-Cahn and Phase-Field elements
