+++ test/test_phase_field_element.py
@@ -1,0 +1,99 @@
+import pytest
+import numpy as np
+import pyvista as pv
+import os
+import dolfinx
+import mpi4py
+from phasefieldx.Element.Phase_Field.Input import Input
+from phasefieldx.Element.Phase_Field.solver.solver import solve
+from phasefieldx.Boundary.boundary_conditions import bc_phi, get_ds_bound_from_marker
+from phasefieldx.PostProcessing.ReferenceResult import AllResults
+
+@pytest.mark.parametrize("dimension", ["1d", "2d", "3d"])
+def test_phase_field_simulation(dimension):
+    """
+    Test phase-field simulation for 1D, 2D, and 3D cases and compare results with theoretical values.
+    """
+    # Set parameters for each dimension
+    lx, ly, lz = 5.0, 1.0, 1.0
+    divx, divy, divz = 200, 1, 1
+
+    # Create mesh depending on the dimension
+    if dimension == "1d":
+        msh = dolfinx.mesh.create_interval(mpi4py.MPI.COMM_WORLD,
+                                           divx,
+                                           np.array([0.0,  lx]))
+    elif dimension == "2d":
+        msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
+                                            [np.array([0.0, 0.0]),
+                                            np.array([lx, ly])],
+                                            [divx, divy],
+                                            cell_type=dolfinx.mesh.CellType.quadrilateral)
+    elif dimension == "3d":
+        msh = dolfinx.mesh.create_box(mpi4py.MPI.COMM_WORLD, 
+                                      [np.array([0.0, 0.0, 0.0]),
+                                      np.array([lx, ly, lz])],
+                                      [divx, divy, divz], 
+                                      cell_type=dolfinx.mesh.CellType.hexahedron)
+
+    # Set phase-field input
+    Data = Input(l=4.0, save_solution_xdmf=False, save_solution_vtu=True, results_folder_name="test_results")
+
+    # Define boundary condition and function space
+    def left(x):
+        return np.equal(x[0], 0)
+
+    fdim = msh.topology.dim - 1
+    left_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, left)
+    ds_left = get_ds_bound_from_marker(left_facet_marker, msh, fdim)
+    ds_list = np.array([[ds_left, "left"]])
+    
+    V_phi = dolfinx.fem.functionspace(msh, ("Lagrange", 1))
+    bc_left = bc_phi(left_facet_marker, V_phi, fdim, value=1.0)
+    bcs_list_phi = [bc_left]
+
+    # Solve the problem
+    solve(Data,
+          msh,
+          1.0,
+          V_phi,
+          bcs_list_phi,
+          None,
+          None,
+          ds_list,
+          1.0,
+          path=None,
+          quadrature_degree=2)
+
+    # Load results using PyVista
+    #pv.start_xvfb()
+    file_vtu = pv.read(os.path.join(Data.results_folder_name, "paraview-solutions_vtu", "phasefieldx_p0_000000.vtu"))
+
+    # Theoretical phase-field values along the x-axis
+    xt = file_vtu.points[:,0]#np.linspace(-lx, lx, divx)
+    phi_theory = np.exp(-abs(xt)/Data.l) + 1/(np.exp(2*lx/Data.l)+1) * 2* np.sinh(np.abs(xt)/Data.l)
+
+    # Compare numerical results with theoretical values (tolerance is set to allow for slight deviations)
[…]
