### 2024-10-10 7ca6363
Fix division by zero error in L2 norm calculation by adding a check for zero normalization factor
