+++ docs/environment.yml
@@ -8,7 +8,7 @@ dependencies:
   - numpy
   - vtk
   - scooby>=0.5.1
-  - fenics-dolfinx
+  - fenics-dolfinx==0.8.0
   - meshio
   - matplotlib
   - imageio>=2.5.0
