### 2024-10-10 7052364
Added `quadrature_degree` parameter
