### 2024-10-10 34c9c9f
Refactored the `solve` function to focus on the Allen-Cahn equation solver. Changes include:
- Added detailed docstring to describe function parameters, return values, and usage.
- Introduced a `quadrature_degree` parameter for flexibility in integration accuracy.
- Modified energy formulation to align with Allen-Cahn dynamics.
- Updated simulation logging and result saving to match the new problem formulation.
- Ensured compatibility with Paraview output in both XDMF and VTU formats.
