+++ src/phasefieldx/Element/Allen_Cahn/solver/solver.py
@@ -1,5 +1,8 @@
 '''
-Phase-field Problem
+Solver: Allen-Cahn
+==================
+
+
 '''
 
 # Libraries ############################################################
@@ -8,15 +11,13 @@ import os
 import time
 import dolfinx
 import ufl
-from dolfinx.fem.petsc import NonlinearProblem
-
 
 from phasefieldx.files import prepare_simulation, append_results_to_file
 from phasefieldx.solvers.newton import NewtonSolver
 from phasefieldx.Logger.library_versions import set_logger, log_library_versions, log_system_info, log_end_analysis, log_model_information
 
-def solve(path, 
-          Data, 
+
+def solve(Data, 
           msh, 
           final_time, 
           V_phi, 
@@ -25,8 +26,59 @@ def solve(path,
           update_loading=None, 
           initial_condition=None,
           ds_bound=None,
-          dt = 1.0):
+          dt = 1.0,
+          path = None,
+          quadrature_degree = 2):
+    """
+    Solver for the Allen-Cahn equation.
+
+    Parameters
+    ----------
+    Data : phasefieldx.Data
+        Data object containing simulation parameters.
+    msh : dolfinx.cpp.mesh.Mesh
+        Mesh object defining the computational domain.
+    final_time : float
+        Final pseudo time for the simulation.
+    V_phi : dolfinx.fem.FunctionSpace
+        Function space for the phase-field variable phi.
+    bc_list_phi : list of dolfinx.fem.DirichletBC, optional
+        List of Dirichlet boundary conditions for phi (default is []).
+    update_boundary_conditions : function, optional
+        Function to update boundary conditions based on time (default is None).
+    update_loading : function, optional
+        Function to update external loading conditions based on time (default is None).
+    initial_condition : dolfinx.fem.Function, optional
+        Initial condition for phi (default is None).
+    ds_bound : numpy.ndarray, optional
+        Array containing boundary descriptions for reaction forces (default is None).
+    dt : float, optional
+        Time step size (default is 1.0).
+    path : str, optional
+        Path to store simulation results (default is current working directory).
+
+    Returns
+    -------
+    None
+
+    Notes
+    -----
+    This function initializes and solves the Allen-Cahn equation for the phase-field variable phi,
+    which represents a free energy minimization problem. It uses a Newton-type solver to update phi
+    over the specified time period. Simulation progress is logged, and results are saved using
+    Paraview-compatible formats.
+
+    Examples
+    --------
+    # Initialize Data, msh, V_phi, and optionally update functions
+    solve(Data, msh, 10.0, V_phi, bc_list_phi, update_boundary_conditions, update_loading)
+
+    # This will simulate the free energy minimization problem using Allen-Cahn equation,
+    # saving results in the specified directory.
+    """
     
+    if path==None:
+        path = os.getcwd()
     # Common #############################################################
     ######################################################################
     result_folder_name = Data.results_folder_name
@@ -57,13 +109,14 @@ def solve(path,
         
     f = 0.25*(phi**2 -1)**2
     dfdc = ufl.diff(f,phi)
-    #metadata = {"quadrature_degree": 8}
+    metadata = {"quadrature_degree": quadrature_degree}
[…]
