### 2024-10-10 95e2184
Update version to 0.0.3
