+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "phasefieldx"
-version = "0.0.2"
+version = "0.0.3"
 authors = [
   { name="Miguel Castillon", email="phasefieldx@gmail.com" },
 ]
