### 2024-10-11 a286a16
style: format all source code to comply with PEP8 guidelines
