+++ src/phasefieldx/Element/Phase_Field/solver/__init__.py
@@ -1,4 +1,3 @@
 # phasefieldx/Element/Phase_Field/solver/__init__.py
 
 from .solver import *
-
