+++ src/phasefieldx/Element/Phase_Field_Fracture/solver/__init__.py
@@ -3,7 +3,8 @@
 from .solver import *
 
 
-# Optionally, you can specify which symbols to export when using 'from phasefieldx import *'
+# Optionally, you can specify which symbols to export when using 'from
+# phasefieldx import *'
 __all__ = [
     'solver',
 ]
