### 2024-10-11 55440ae
style: format examples code to comply with PEP8 guidelines
