+++ examples/PhaseFieldFracture/plot_1702.py
@@ -6,19 +6,19 @@ One Element tension anisotropic (spectral)
 
 .. code-block::
 
-   #          u /\        /\  
-   #            ||        ||      
+   #          u /\\        /\
+   #            ||        ||
    #     (0, 1) *----------* (1, 1)
-   #            |          | 
    #            |          |
    #            |          |
    #            |          |
-   #            |          | 
+   #            |          |
+   #            |          |
    #     (0, 0) *----------* (0, 1)
-   #            /_\       /_\       
+   #            /_\\       /_\
    #     |Y    ////       ///
-   #     |                
-   #     *---X           
+   #     |
+   #     *---X
 
 
 .. _table_properties_label:
@@ -39,16 +39,16 @@ One Element tension anisotropic (spectral)
 
 In this example, we study a simple model formed by a single element with four nodes and dimensions of 1x1 mm. The bottom nodes are constrained in both directions, and the top nodes can slide vertically.
 
-The Young's modulus, Poisson's ratio, and the critical energy release rate are given in the table :ref:`Properties <table_properties_label>`. Young's modulus $E$ and Poisson's ratio $\nu$ can be represented with the Lamé parameters as: $\lambda=\frac{E\nu}{(1+\nu)(1-2\nu)}$; $\mu=\frac{E}{2(1+\nu)}$.
+The Young's modulus, Poisson's ratio, and the critical energy release rate are given in the table :ref:`Properties <table_properties_label>`. Young's modulus $E$ and Poisson's ratio $\nu$ can be represented with the Lamé parameters as: $\\lambda=\frac{E\nu}{(1+\nu)(1-2\nu)}$; $\\mu=\frac{E}{2(1+\nu)}$.
 
-In this case, due to the discretization, it is possible to obtain an analytical solution for the anisotropic model by solving $\phi$ from the given equations. The term $|\nabla \phi|^2$ vanishes due to the discretization as explained by Molnar \cite{MOLNAR201727} and Miehe \cite{Miehe1} in the appendix.
+In this case, due to the discretization, it is possible to obtain an analytical solution for the anisotropic model by solving $\\phi$ from the given equations. The term $|\nabla \\phi|^2$ vanishes due to the discretization as explained by Molnar \\cite{MOLNAR201727} and Miehe \\cite{Miehe1} in the appendix.
 
 .. math::
-   \phi = \frac{2 \psi_a}{\frac{G_c}{l}+2\psi_a}=\frac{2 H}{\frac{G_c}{l}+2H} 
-   
-.. math:: 
-    \sigma_y = \sigma_{a}(1-\phi)^2 + \sigma_{b}
-    
+   \\phi = \frac{2 \\psi_a}{\frac{G_c}{l}+2\\psi_a}=\frac{2 H}{\frac{G_c}{l}+2H}
+
+.. math::
+    \\sigma_y = \\sigma_{a}(1-\\phi)^2 + \\sigma_{b}
+
 """
 
 ###############################################################################
@@ -58,7 +58,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import pyvista as pv
 import dolfinx
-import mpi4py 
+import mpi4py
 import petsc4py
 import os
 
@@ -75,17 +75,17 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 # Parameters Definition
 # ---------------------
 # Here, we define the class containing the input parameters. The material parameters are:
-# Young's modulus $E = 210$, $\text{kN/mm}^2$, Poisson's ratio $\nu = 0.3$, 
[…]
