+++ examples/PhaseFieldFracture/plot_1700.py
@@ -6,19 +6,19 @@ One Element tension Isotropic
 
 .. code-block::
 
-   #          u /\        /\  
-   #            ||        ||      
+   #          u /\        /\
+   #            ||        ||
    #     (0, 1) *----------* (1, 1)
-   #            |          | 
    #            |          |
    #            |          |
    #            |          |
-   #            |          | 
+   #            |          |
+   #            |          |
    #     (0, 0) *----------* (0, 1)
-   #            /_\       /_\       
+   #            /_\       /_\
    #     |Y    ////       ///
-   #     |                
-   #     *---X           
+   #     |
+   #     *---X
 
 
 .. _table_properties_label:
@@ -44,11 +44,11 @@ The Young's modulus, Poisson's ratio, and the critical energy release rate are g
 In this case, due to the discretization, it is possible to obtain an analytical solution for the isotropic model by solving $\phi$ from the given equations. The term $|\nabla \phi|^2$ vanishes due to the discretization as explained by Molnar \cite{MOLNAR201727} and Miehe \cite{Miehe1} in the appendix.
 
 .. math::
-   \phi = \frac{2 \psi_a}{\frac{G_c}{l}+2\psi_a}=\frac{2 H}{\frac{G_c}{l}+2H} 
-   
-.. math:: 
+   \phi = \frac{2 \psi_a}{\frac{G_c}{l}+2\psi_a}=\frac{2 H}{\frac{G_c}{l}+2H}
+
+.. math::
     \sigma_y = \sigma_{a}(1-\phi)^2 + \sigma_{b}
-    
+
 """
 
 ###############################################################################
@@ -58,7 +58,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import pyvista as pv
 import dolfinx
-import mpi4py 
+import mpi4py
 import petsc4py
 import os
 
@@ -76,17 +76,17 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 # Parameters Definition
 # ---------------------
 # Here, we define the class containing the input parameters. The material parameters are:
-# Young's modulus $E = 210$, $\text{kN/mm}^2$, Poisson's ratio $\nu = 0.3$, 
-# critical energy release rate $G_c = 0.005$, $\text{kN/mm}^2$, and length scale parameter $l = 0.1$, $\text{mm}$. 
+# Young's modulus $E = 210$, $\text{kN/mm}^2$, Poisson's ratio $\nu = 0.3$,
+# critical energy release rate $G_c = 0.005$, $\text{kN/mm}^2$, and length scale parameter $l = 0.1$, $\text{mm}$.
 # We consider isotropic degradation where all the energy is degraded and irreversibility as proposed by Miehe.
 Data = Input(E=210.0,   # young modulus
              nu=0.3,    # poisson
              Gc=0.005,  # critical energy release rate
              l=0.1,     # lenght scale parameter
-             degradation="isotropic", # "isotropic" "anisotropic"
+             degradation="isotropic",  # "isotropic" "anisotropic"
              split_energy="no",       # "spectral" "deviatoric"
              degradation_function="quadratic",
-             irreversibility="miehe", # "miehe"
+             irreversibility="miehe",  # "miehe"
              fatigue=False,
              fatigue_degradation_function="asymptotic",
              fatigue_val=0.05625,
@@ -99,7 +99,6 @@ Data = Input(E=210.0,   # young modulus
              results_folder_name="1700_One_element_isotropic_tension")
 
 
-
[…]
