+++ examples/GmshGeoFiles/plot_9102.py
@@ -8,14 +8,14 @@
 ---------
 .. include::  ../../../../examples/GmshGeoFiles/9102_SingleNotchedShearTest/file.geo
    :literal:
-   
+
 """
 
 ###############################################################################
 # Mesh Visualization
 # ------------------
-# The purpose of this code is to visualize the mesh. The mesh is generated from 
-# the .geo file and saved as output_mesh_for_view.vtu. It is then loaded and 
+# The purpose of this code is to visualize the mesh. The mesh is generated from
+# the .geo file and saved as output_mesh_for_view.vtu. It is then loaded and
 # visualized using PyVista.
 
 import os
@@ -44,4 +44,4 @@ print(f"Mesh successfully written to {vtu_file}")
 
 pv.start_xvfb()
 file_vtu = pv.read(vtu_file)
-file_vtu.plot(cpos='xy',color='white', show_edges=True)
+file_vtu.plot(cpos='xy', color='white', show_edges=True)

[…]
