### 2024-10-11 dec1f5d
chore: update project metadata and add license file to pyproject.toml
