+++ test/Element/Phase_Field/test_phase_field_element.py
@@ -4,15 +4,25 @@ import pyvista as pv
 import os
 import dolfinx
 import mpi4py
+
+# Import necessary components from the phasefieldx package
 from phasefieldx.Element.Phase_Field.Input import Input
 from phasefieldx.Element.Phase_Field.solver.solver import solve
 from phasefieldx.Boundary.boundary_conditions import bc_phi, get_ds_bound_from_marker
 from phasefieldx.PostProcessing.ReferenceResult import AllResults
 
+
 @pytest.mark.parametrize("dimension", ["1d", "2d", "3d"])
 def test_phase_field_simulation(dimension):
     """
-    Test phase-field simulation for 1D, 2D, and 3D cases and compare results with theoretical values.
+    Test phase-field simulation for various dimensions (1D, 2D, 3D).
+
+    Parameters:
+    dimension (str): The dimension of the phase-field simulation. Options are '1d', '2d', '3d'.
+
+    Asserts:
+    - Phase-field values match theoretical predictions.
+    - Computed energy values match theoretical values within a specified tolerance.
     """
     # Set parameters for each dimension
     lx, ly, lz = 5.0, 1.0, 1.0
@@ -22,22 +32,23 @@ def test_phase_field_simulation(dimension):
     if dimension == "1d":
         msh = dolfinx.mesh.create_interval(mpi4py.MPI.COMM_WORLD,
                                            divx,
-                                           np.array([0.0,  lx]))
+                                           np.array([0.0, lx]))
     elif dimension == "2d":
         msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
                                             [np.array([0.0, 0.0]),
-                                            np.array([lx, ly])],
+                                             np.array([lx, ly])],
                                             [divx, divy],
                                             cell_type=dolfinx.mesh.CellType.quadrilateral)
     elif dimension == "3d":
-        msh = dolfinx.mesh.create_box(mpi4py.MPI.COMM_WORLD, 
+        msh = dolfinx.mesh.create_box(mpi4py.MPI.COMM_WORLD,
                                       [np.array([0.0, 0.0, 0.0]),
-                                      np.array([lx, ly, lz])],
-                                      [divx, divy, divz], 
+                                       np.array([lx, ly, lz])],
+                                      [divx, divy, divz],
                                       cell_type=dolfinx.mesh.CellType.hexahedron)
 
     # Set phase-field input
-    Data = Input(l=4.0, save_solution_xdmf=False, save_solution_vtu=True, results_folder_name="test_results")
+    Data = Input(l=4.0, save_solution_xdmf=False,
+                 save_solution_vtu=True, results_folder_name="test_results")
 
     # Define boundary condition and function space
     def left(x):
@@ -47,7 +58,7 @@ def test_phase_field_simulation(dimension):
     left_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, left)
     ds_left = get_ds_bound_from_marker(left_facet_marker, msh, fdim)
     ds_list = np.array([[ds_left, "left"]])
-    
+
     V_phi = dolfinx.fem.functionspace(msh, ("Lagrange", 1))
     bc_left = bc_phi(left_facet_marker, V_phi, fdim, value=1.0)
     bcs_list_phi = [bc_left]
[…]
