+++ test/Element/Phase_Field_Fracture/test_phase_field_fracture_element_anisotropic.py
@@ -1,0 +1,156 @@
+import pytest
+import numpy as np
+import pyvista as pv
+import os
+import dolfinx
+import mpi4py
+import petsc4py
+
+# Import necessary components from the phasefieldx package
+from phasefieldx.Element.Phase_Field_Fracture.Input import Input
+from phasefieldx.Element.Phase_Field_Fracture.solver.solver import solve
+from phasefieldx.Boundary.boundary_conditions import bc_xy, get_ds_bound_from_marker
+from phasefieldx.PostProcessing.ReferenceResult import AllResults
+from phasefieldx.Math import macaulay_bracket_positive, macaulay_bracket_negative
+
+
+@pytest.mark.parametrize("split_energy_i", ["spectral"])
+def test_phase_field_simulation_anisotropic(split_energy_i):
+    """
+    Test phase-field simulation under anisotropic conditions and compare results with theoretical values.
+    """
+    Data = Input(E=210.0,   # young modulus
+                 nu=0.3,    # poisson
+                 Gc=0.005,  # critical energy release rate
+                 l=0.1,     # lenght scale parameter
+                 degradation="anisotropic",  # "isotropic" "anisotropic"
+                 split_energy=split_energy_i,       # "spectral" "deviatoric"
+                 degradation_function="quadratic",
+                 irreversibility="no",  # "miehe"
+                 fatigue=False,
+                 fatigue_degradation_function="asymptotic",
+                 fatigue_val=0.05625,
+                 k=0.0,
+                 min_stagger_iter=2,
+                 max_stagger_iter=500,
+                 stagger_error_tol=1e-8,
+                 save_solution_xdmf=False,
+                 save_solution_vtu=True,
+                 results_folder_name="1700_One_element_isotropic_tension_test_anisotropic")
+
+    ###############################################################################
+    # Mesh Definition
+    # ---------------
+    # We create a 1x1 mm quadrilateral mesh with a single element using Dolfinx.
+    msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
+                                        [np.array([0, 0]),
+                                         np.array([1, 1])],
+                                        [1, 1],
+                                        cell_type=dolfinx.mesh.CellType.quadrilateral)
+
+    def bottom(x):
+        return np.isclose(x[1], 0)
+
+    def top(x):
+        return np.isclose(x[1], 1)
+
+    fdim = msh.topology.dim - 1
+
+    bottom_facet_marker = dolfinx.mesh.locate_entities_boundary(
+        msh, fdim, bottom)
+    top_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, top)
+
+    ds_bottom = get_ds_bound_from_marker(top_facet_marker, msh, fdim)
+    ds_top = get_ds_bound_from_marker(top_facet_marker, msh, fdim)
+
+    ds_list = np.array([
+        [ds_bottom, "bottom"],
+        [ds_top, "top"]
+    ])
+
+    ###############################################################################
+    # Function Space Definition
+    # -------------------------
+    # Define function spaces for displacement and phase-field using Lagrange elements.
+    V_u = dolfinx.fem.functionspace(msh, ("Lagrange", 1, (msh.geometry.dim, )))
+    V_phi = dolfinx.fem.functionspace(msh, ("Lagrange", 1))
+
+    ###############################################################################
+    # Boundary Conditions
+    # -------------------
[…]

[…]
