### 2024-10-11 a495818
Make version dynamic and add metadata for documentation and keywords
- Set project version to be dynamically managed from `__init__.py` file using Hatch.
- Added keywords related to phase-field simulations (fracture, fatigue, solidification, etc.).
- Updated project URLs to include a link to the official documentation.
