### 2024-10-14 e3ec618
Add GitHub Actions workflow for building JOSS paper draft
