+++ .github/workflows/build_joss_paper.yml
@@ -1,0 +1,25 @@
+name: Build JOSS paper
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  paper:
+    runs-on: ubuntu-latest
+    name: Paper Draft
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Build draft PDF
+        uses: openjournals/openjournals-draft-action@master
+        with:
+          journal: joss
+          paper-path: ./joss-paper/paper.md
+
+      - name: Upload
+        uses: actions/upload-artifact@v4
+        with:
+          name: paper
+          path: ./joss-paper/paper.pdf
