### 2024-10-14 6af8559
Add three new tests to improve coverage
