+++ test/test_version.py
@@ -1,0 +1,5 @@
+import phasefieldx
+
+
+def test_version():
+    assert phasefieldx.__version__ is not None
