### 2024-10-14 a8b2a9b
Refactor find_files_with_extension function for clarity and correctness
