### 2024-10-14 af941ea
Update JOSS paper workflow paths
