+++ .github/workflows/build_joss_paper.yml
@@ -16,10 +16,10 @@ jobs:
         uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
-          paper-path: ./joss-paper/paper.md
+          paper-path: ./joss/paper.md
 
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
           name: paper
-          path: ./joss-paper/paper.pdf
+          path: ./joss/paper.pdf
