### 2024-10-16 4eeee40
Add GitHub Actions workflow for automated testing
