+++ test/test_gmsh_geo_file_processing.py
@@ -18,30 +18,30 @@ def setup_gmsh():
     gmsh.finalize()
 
 
-@pytest.mark.parametrize("geo_file, vtu_file", zip(GEO_FILES, VTU_FILES))
-def test_mesh_generation(geo_file, vtu_file):
-    """Test the mesh generation from .geo files."""
-    # Open the .geo file
-    full_geo_path = os.path.join(FOLDER, geo_file)
-    gmsh.open(full_geo_path)
+# @pytest.mark.parametrize("geo_file, vtu_file", zip(GEO_FILES, VTU_FILES))
+# def test_mesh_generation(geo_file, vtu_file):
+#     """Test the mesh generation from .geo files."""
+#     # Open the .geo file
+#     full_geo_path = os.path.join(FOLDER, geo_file)
+#     gmsh.open(full_geo_path)
 
-    # Generate the mesh (2D example, for 3D use generate(3))
-    gmsh.model.mesh.generate(3)
+#     # Generate the mesh (2D example, for 3D use generate(3))
+#     gmsh.model.mesh.generate(3)
 
-    # Write the mesh to a .vtu file
-    gmsh.write(vtu_file)
+#     # Write the mesh to a .vtu file
+#     gmsh.write(vtu_file)
 
-    # Check if the .vtu file is created
-    assert os.path.isfile(vtu_file), f"{vtu_file} was not created"
+#     # Check if the .vtu file is created
+#     assert os.path.isfile(vtu_file), f"{vtu_file} was not created"
 
-    # Test reading the generated .vtu file with PyVista
+#     # Test reading the generated .vtu file with PyVista
 
-    # Attempt to read the .vtu file
-    mesh = pv.read(vtu_file)
+#     # Attempt to read the .vtu file
+#     mesh = pv.read(vtu_file)
 
-    # Check if the mesh is not empty
-    assert mesh.n_cells > 0, f"The mesh from {vtu_file} is empty or not read properly."
-    assert mesh.n_points > 0, f"The mesh from {vtu_file} has no points."
+#     # Check if the mesh is not empty
+#     assert mesh.n_cells > 0, f"The mesh from {vtu_file} is empty or not read properly."
+#     assert mesh.n_points > 0, f"The mesh from {vtu_file} has no points."
 
-    # Optional: Uncomment to visualize the mesh during manual testing
-    # mesh.plot(cpos='xy', color='white', show_edges=True)  # Uncomment for visualization
+#     # Optional: Uncomment to visualize the mesh during manual testing
+#     # mesh.plot(cpos='xy', color='white', show_edges=True)  # Uncomment for visualization
