+++ requirements.txt
@@ -1,0 +1,21 @@
+enum-tools
+imageio
+imageio-ffmpeg
+jupyterlab
+lxml
+meshio
+mpi4py
+mypy
+mypy-extensions
+numpydoc
+osmnx
+pypandoc
+pytest-sphinx
+scipy
+sympy
+trame
+trame-vtk
+trame-vuetify
+trimesh
+pandas
+gmsh
