### 2024-10-16 CastillonMiguel/phasefieldx#3 @CastillonMiguel
I have added a basic test framework. All tests are located in a new folder called test/, and they are run using pytest. I am now checking some of the models with simple test cases that allow for comparison of the numerical solutions with the analytical ones. Additionally, I have included tests related to general utilities. Furthermore, I have implemented a GitHub workflow that automatically runs all tests whenever a push or pull request is made.
https://github.com/CastillonMiguel/phasefieldx/commit/4eeee4014a533905ec7b7e9908b0a0607f9b1cb1
