### 2024-10-16 5ab9745
Enhance contributing guide: Add section on writing and organizing tests
