### 2024-10-16 CastillonMiguel/phasefieldx#1 @CastillonMiguel
I have added a new section to the contributing guide outlining how to write and organize tests. https://github.com/CastillonMiguel/phasefieldx/commit/5ab97458c4a4b0ef46dfcf2d576c4960667ff3df
