### 2024-10-16 251a378
Add GitHub Actions workflow for publishing to PyPI
