### 2024-10-16 c9fb118
Update version to test GitHub Actions publish workflow
