+++ src/phasefieldx/__init__.py
@@ -1,6 +1,6 @@
 # phasefieldx/__init__.py
 
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 __author__ = "Miguel Castillón"
 __email__ = "phasefieldx@gmail.com"
 __license__ = "MIT"
