+++ README.rst
@@ -72,7 +72,7 @@ Follow these steps to set up your environment:
    
    .. code-block::
    
-      conda install -c conda-forge fenics-dolfinx mpich pyvista pandas
+      conda install -c conda-forge fenics-dolfinx=0.8.0 mpich pyvista pandas
 
 4. Install `gmsh`
    
