### 2024-10-21 3395d76
docs: update installation instructions to include additional required packages
