### 2024-10-23 40c3a36
Update vector call for FEniCSx 0.9.0 compatibility
