+++ src/phasefieldx/Math/projection.py
@@ -49,7 +49,7 @@ def project(e, target_func, bcs=[]):
     solver.getPC().setType("bjacobi")
     # solver.rtol = 1.0e-05
     solver.setOperators(A)
-    solver.solve(b, target_func.vector)
+    solver.solve(b, target_func.x.petsc_vec)
     assert solver.reason > 0
     target_func.x.scatter_forward()
 
