### 2024-10-23 1dc206d
Update FEniCSx dependency to version 0.9.0 for compatibility and new features
