### 2024-10-23 c6f5311
set static version 0.0.5
