### 2024-10-25 fb2b8ea
Refactor environment dependencies in phasefieldx-env
- Reordered dependencies alphabetically for consistency
- Simplified `pip` section to include requirements from `requirements.txt`
- Removed unused or redundant packages
