+++ requirements.txt
@@ -1,21 +1,1 @@
-enum-tools
-imageio
-imageio-ffmpeg
-jupyterlab
-lxml
-meshio
-mpi4py
-mypy
-mypy-extensions
-numpydoc
-osmnx
-pypandoc
-pytest-sphinx
-scipy
-sympy
-trame
-trame-vtk
-trame-vuetify
-trimesh
-pandas
 gmsh
