### 2024-10-30 34a6667
set dynamic versioning and add dependencies
