### 2024-10-30 e10442f
Update installation instructions for conda-forge dependencies
