### 2024-10-30 237c024
Bump version to 0.0.6 in __init__.py
