### 2024-11-04 CastillonMiguel/phasefieldx#5 @finsberg
[JOSS REVIEW] Installation instructions
I belive the installation instructions could be simplified. I would suggest the following.

1. ~Add PhaseFieldx to pypi (Check out a guide [here](https://packaging.python.org/en/latest/tutorials/packaging-projects/) , and see [here](https://packaging.python.org/en/latest/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/) how you can upload a new version with GitHub actions every time you make a new release). The reason for this is that many FEniCSx users already have installed FEniCSx (either from source, conda or docker), and being able to just `pip install phasefieldx` would be convenient.~
2. Publish a docker image with your package (You can see an example [here](https://github.com/finsberg/fenicsx-pulse/pkgs/container/fenicsx-pulse) with a GitHub action to build it [here](https://github.com/finsberg/fenicsx-pulse/blob/main/.github/workflows/docker-image.yml). This is nice because user that pull your image and everything they need.
3. Finally, I would recommend to create a conda package (related to https://github.com/CastillonMiguel/phasefieldx/issues/2), so that you don't need to specify everything the user has to install but rather state these dependencies in you conda package dependencies.

Edit: I saw that you have already added phasefieldx to pypi

This issue is related to the review here: https://github.com/openjournals/joss-reviews/issues/7307
