### 2024-11-04 CastillonMiguel/phasefieldx#6 @finsberg
[JOSS REVIEW] Compare your software with other commonly-used packages 
According to the [Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper), the paper should contain *A description of how this software compares to other commonly-used packages in this research area*. I don't see such a comparison in your paper.

This issue is related to the review here: https://github.com/openjournals/joss-reviews/issues/7307
