### 2024-11-04 CastillonMiguel/phasefieldx#7 @finsberg
[JOSS REVIEW] Improve Summary
According to the [Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper) the paper should give *A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.* My opinion is that the current summary is a bit too short, and require a descent level of expertise. I think it could be improved by showing some examples of where relevant applications arises and and what a phase field model is. Are there some relevant equations? How does this compare to "regular" solid mechanics.

This issue is related to the review here: https://github.com/openjournals/joss-reviews/issues/7307
