### 2024-11-04 CastillonMiguel/phasefieldx#8 @finsberg
[JOSS REVIEW] Improve Statement of Need
According to the [Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#the-joss-paper), the paper should contain *A description of how this software compares to other commonly-used packages in this research area*. I think you described that it is important with a good community, but I don't really see a good description of why we need `PhaseFieldx`. For example, why cannot users just use FEniCSx directly. Why do we need a separate package for this? I would recommend to checkout the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) to get some inspiration to what you can write in this section.

This issue is related to the review here: https://github.com/openjournals/joss-reviews/issues/7307
