### 2024-11-04 CastillonMiguel/phasefieldx#9 @finsberg
[JOSS REVIEW] Improve description in examples
I must first say that I really like the way you have created the examples. It is really cool with both the pyvista integration and the "Download buttons" at the end. Really nice job!

I think however that some of the examples could be improved by providing some more text explaining the different examples, preferable with the necessary mathematical background and references to papers. For example in [Displacement control](https://phasefieldx.readthedocs.io/en/latest/auto_examples/Elasticity/plot_1100.html#sphx-glr-auto-examples-elasticity-plot-1100-py) there are not really any text explaining what this example is about, and it seems to be some headings and text that are copy-pasted from other examples.

Instead of creating a separate issue for each example I create this one as a general issue for all of them, and then you can go through the examples and make sure that each example reads as a coherent piece of text.

This issue is related to the review here: https://github.com/openjournals/joss-reviews/issues/7307
