### 2024-11-04 CastillonMiguel/phasefieldx#10 @finsberg
[JOSS REVIEW] Elaborate on applications
In your paper you mention some relevant applications. I think it would be good to give some more details about the applications and perhaps also add some figures showing the results of the simualtions using your software.

This issue is related to the review here: https://github.com/openjournals/joss-reviews/issues/7307
