### 2024-11-05 CastillonMiguel/phasefieldx#7 @CastillonMiguel
Thank you for the constructive feedback. I’ll expand the summary to make it more accessible, including examples of applications for phase field models and key equations to clarify its functionality. I plan to complete this by the end of the week.
