### 2024-11-05 CastillonMiguel/phasefieldx#5 @CastillonMiguel
I’ll check on publishing a Docker image, as this could indeed simplify setup for users.

Regarding Conda-Forge, I’m currently working on this. The package is on Conda-Forge, but it’s not fully functional yet due to some issues I need to resolve. It’s taking a bit more time than anticipated, but I’m actively addressing these challenges to ensure it works smoothly.
https://github.com/conda-forge/phasefieldx-feedstock
https://conda-forge.org/packages/?
