### 2024-11-07 487b67a
Update JOSS paper with latest changes and improvements
