+++ joss/paper.bib
@@ -53,6 +53,19 @@ url = {https://www.sciencedirect.com/science/article/pii/S0045782510001283},
 author    = {Miehe, C. and Hofacker, M. and Welschinger, F.},
 }
 
+@article{FrancfortMarigo1998,
+title = {Revisiting brittle fracture as an energy minimization problem},
+journal = {Journal of the Mechanics and Physics of Solids},
+volume = {46},
+number = {8},
+pages = {1319-1342},
+year = {1998},
+issn = {0022-5096},
+doi = {10.1016/S0022-5096(98)00034-9},
+url = {https://www.sciencedirect.com/science/article/pii/S0022509698000349},
+author = {Francfort, G.A. and Marigo, J.-J.},
+}
+
 @article{Amor2009,
 title = {Regularized formulation of the variational brittle fracture with unilateral contact: Numerical experiments},
 journal = {Journal of the Mechanics and Physics of Solids},
@@ -69,7 +82,6 @@ keywords = {Fracture, Energy methods, Free-discontinuity problems, Contact mecha
 
 
 @Article{Ambati2015,
-author    = {Ambati, M. and Gerasimov, T. and De Lorenzis, L.},
 title={A review on phase-field models of brittle fracture and a new fast hybrid formulation},
 journal={Computational Mechanics},
 year={2015},
@@ -80,7 +92,8 @@ number={2},
 pages={383-405},
 issn={1432-0924},
 doi={10.1007/s00466-014-1109-y},
-url={https://doi.org/10.1007/s00466-014-1109-y}
+url={https://doi.org/10.1007/s00466-014-1109-y},
+author = {Ambati, M. and Gerasimov, T. and De Lorenzis, L.},
 }
 
 @article{Carrara2020,
