+++ joss/paper.md
@@ -3,6 +3,7 @@ title: "PhaseFieldX: An Open-Source Framework for Advanced Phase-Field Simulatio
 tags:
   - Fenicsx
   - Phase field
+  - Fatigue
   - Fracture
   - Finite element
 authors:
@@ -20,27 +21,76 @@ bibliography: paper.bib
 
 # Summary
 
-The **PhaseFieldX** project is designed to simulate and analyze material behavior using phase-field models, which provide a continuous approximation of interfaces, phase boundaries, and discontinuities such as cracks. Leveraging the robust capabilities of *FEniCSx* [@BarattaEtal2023; @ScroggsEtal2022; @BasixJoss; @AlnaesEtal2014], a renowned finite element framework for solving partial differential equations, this project facilitates efficient and precise numerical simulations. It supports a wide range of applications, including phase-field fracture, solidification, and other complex material phenomena, making it an invaluable resource for researchers and engineers in materials science.
+The **PhaseFieldX** project is designed to simulate and analyze material behavior using phase-field models, which provide a continuous approximation for interfaces, phase boundaries, and discontinuities such as cracks. Leveraging the robust capabilities of *FEniCSx* [@BarattaEtal2023; @ScroggsEtal2022; @BasixJoss; @AlnaesEtal2014], a renowned finite element framework for solving partial differential equations, this project enables efficient and accurate numerical simulations related to phase-field models. It supports a broad range of applications, including phase-field fracture, solidification, and other complex material phenomena, making it an invaluable resource for researchers and engineers in materials science.
 
-![Logo](./images/logo_name.png){height="80pt"}
+![PhaseFieldX logo.\label{fig:logo}](./images/logo_name.png){height="100pt"}
+
+- **Phase-Field Fracture**:
+
+The phase-field fracture model was extended by [@Miehe2010] to incorporate a regularized crack field within the variational framework of Griffith’s fracture theory, as developed by [@FrancfortMarigo1998]. This model provides a robust and versatile approach for simulating fracture and material failure, representing cracks implicitly through a continuous phase-field variable rather than requiring explicit crack tracking. This implicit representation simplifies numerical implementation and naturally enables the simulation of complex crack patterns, including branching, merging, and interactions among multiple cracks.
+
+A diffuse variable, the “phase-field” $\phi$, distinguishes between intact ($\phi = 0$) and fractured ($\phi = 1$) regions, with a smooth transition between them. This setup allows cracks to evolve based on energy minimization principles, eliminating the need for explicit crack-path tracking. The functional is given by \autoref{eq:phase_field_fracture_functional},
+\begin{equation}\label{eq:phase_field_fracture_functional}
+\mathcal{E}(\boldsymbol u, \phi) = \int_\Omega g(\phi)  \psi(\epsilon(\boldsymbol u)) \mathrm{d}\Omega + G_c \int_\Omega \left( \frac{1}{2l}\phi^2 + \frac{l}{2} |\nabla \phi|^2 \right) \mathrm{d}\Omega - E_{\text{ext}}[\boldsymbol u].
+\end{equation}
[…]
