### 2024-11-07 CastillonMiguel/phasefieldx#7 @CastillonMiguel
Expanded summary with examples of phase-field model applications and key equations for clarity. Ready for review

https://github.com/CastillonMiguel/phasefieldx/commit/487b67af6c257d0bf032dd8f01234dbc184ac4e1
