### 2024-11-07 CastillonMiguel/phasefieldx#8 @CastillonMiguel
The changes have been committed and are now reflected in the updated version of the paper.

https://github.com/CastillonMiguel/phasefieldx/commit/487b67af6c257d0bf032dd8f01234dbc184ac4e1
