### 2024-11-08 e79cd38
Update comments and explanations in PhaseField examples
