+++ examples/PhaseField/plot_2000.py
@@ -4,8 +4,11 @@
 Crack surface density functional
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-In the following example, we consider the boundary value problem of the phase-field model, which represents the crack surface density functional, thus providing a continuous approximation of the discontinuous crack  (:ref:`theory_phase_field`). Due to the symmetry of the problem, only the left half of the bar is considered. Therefore, a boundary condition is applied at the left end of this half bar, as illustrated in the following diagrams.
-For a one-dimensional simulation, the boundary condition $\\phi=1$ is set at the point located on the left end.
+In this example, we examine the boundary value problem related to the representation of a crack surface. This approach models the crack surface density functional, offering a continuous approximation of an otherwise discontinuous crack surface (:ref:`theory_phase_field`).
+
+Due to the symmetry of the problem, we consider only the left half of the bar. A boundary condition is therefore applied at the left end of this half-bar, as shown in the diagrams below.
+
+For a one-dimensional simulation, the boundary condition $\phi=1$ is imposed at the left end of this segment.
 
 .. code-block::
 
@@ -75,56 +78,116 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 ###############################################################################
 # Parameters definition
 # ---------------------
-# A length scale parameter of  $l = 4$  is defined. The phase field is saved in a VTU file.
-# All the results are stored in a folder named results_folder_name = name.
-Data = Input(l=4.0,
-             save_solution_xdmf=False,
-             save_solution_vtu=True,
-             results_folder_name="2000_General")
+# First, we define an input class, which contains all the parameters needed for the setup 
+# and results of the simulation.
+#
+# The first term, $l$, specifies the length scale parameter for the problem, with $l = 4.0$.
+#
+# The next two options, `save_solution_xdmf` and `save_solution_vtu`, determine the file format 
+# used to save the phase-field results (.xdmf or .vtu), which can then be visualized using 
+# tools like ParaView or pvista. Both parameters are boolean values (`True` or `False`). 
+# In this case, we set `save_solution_vtu=True` to save the phase-field results in .vtu format.
+#
+# Lastly, `results_folder_name` specifies the name of the folder where all results 
+# and log information will be saved. If the folder does not exist, `phasefieldx` will create it. 
+# However, if the folder already exists, any previous data in it will be removed, and a new 
+# empty folder will be created in its place.
+#
+Data = Input(
+    l=4.0,
+    save_solution_xdmf=False,
+    save_solution_vtu=True,
+    results_folder_name="2000_General"
+)
+
 
 
 ###############################################################################
-# Mesh definition
+# Mesh Definition
 # ---------------
+# We define the mesh parameters and set up a two-dimensional simulation. This setup 
+# supports various dimensions (‘1D’, ‘2D’, or ‘3D’) by creating a mesh that consists 
+# of either a line, rectangle, or box, depending on the selected dimension, with 
+# corresponding line, quadrilateral, or hexahedral elements.
+#
+# - `divx`, `divy`, and `divz` specify the number of divisions along the x, y, and z 
+#   axes, respectively. Here, `divx=100`, `divy=1`, and `divz=1` are set to divide 
+#   the x-axis primarily, as needed for a 2D or 1D mesh.
+#
[…]
