+++ examples/PhaseField/plot_2001.py
@@ -4,7 +4,9 @@
 Representation of a Cracked Plate
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-In the following example, we consider the boundary value problem of the phase-field model (:ref:`theory_phase_field`). A cracked plate will be represented using the phase field variable, following [Miehe]_, with the boundary condition \\phi=1 and different length scale values. Due to the symmetry of the problem, only one half of the plate will be considered. The results will be shown for all the models by applying the reflection of the solution.”
+In this example, we solve a boundary value problem for a cracked plate using the phase-field model (:ref:`theory_phase_field`). The crack is represented by the phase field variable, following the approach presented by [Miehe]_. We apply a boundary condition of $\phi = 1$, representing the broken part of the material, and simulate the crack representation with different length scale values.
+
+To exploit symmetry, only half of the plate is simulated, and we reflect the results to show the full solution. This allows us to reduce computational cost while capturing the entire domain's behavior.
 
 .. code-block::
 
@@ -44,11 +46,16 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 
 
 ###############################################################################
-# Parameters definition
+# Parameter Definition
 # ---------------------
-# Length scale parameters of $l_{10} = 10$, $l_{05} = 0.5$, and $l_{01} = 0.1$
-# are defined. The phase field is saved in a VTU file. All the results are
-# stored in a folder named results_folder_name = name.
+# In this section, we define various length scale parameters, denoted as $l$, 
+# which will be used in the simulations. 
+# We will run three simulations with different values of $l$—specifically, $l=1.0$, $l=0.25$, and $l=0.05$—which will correspond to three distinct input classes. 
+# These length scale parameters control the smoothness of the crack surface:
+# smaller values of $l$ lead to sharper cracks, while larger values result in smoother cracks.
+# The phase-field results will be saved in VTU format for subsequent visualization.
+# For further details about the input class, please refer to :ref:`ref_2000`.
+
 Data1 = Input(l=1.0,
               save_solution_xdmf=False,
               save_solution_vtu=True,
@@ -68,12 +75,15 @@ Data005 = Input(l=0.05,
 ###############################################################################
 # Mesh definition
 # ---------------
+# We define the mesh for the simulations. The mesh covers a rectangular domain
+# with dimensions lx=1.0 and ly=0.5. The mesh is subdivided into `divx` and `divy`
+# divisions along the x and y axes, respectively. In this case, 100 divisions 
+# are made in the x-direction, and 50 in the y-direction, giving us sufficient 
+# resolution for the crack representation.
 divx, divy = 100, 50
 lx, ly = 1.0, 0.5
 
-
-###############################################################################
-# A two-dimensional simulation is considered
+# Create a 2D mesh using the defined parameters.
 msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
                                     [np.array([0, 0]),
                                      np.array([lx, ly])],
@@ -81,16 +91,36 @@ msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
                                     cell_type=dolfinx.mesh.CellType.quadrilateral)
 
 
-###############################################################################
-# The bottom part is denoted and shown to impose the boundary conditions
+# Bottom boundary Identification
+# ------------------------------
+# This function identifies points on the bottom side of the domain where the boundary 
+# condition will be applied. Specifically, it returns `True` for points where `y=0` 
[…]
