### 2024-11-08 CastillonMiguel/phasefieldx#6 @finsberg
Thanks. This look much better :)  However, I still think you need to write a little bit about how your software compares with other commonly-used packages. Here are some that I found when searching online

- https://github.com/prisms-center/phaseField
- https://github.com/SoftSimu/SymPhas
- https://github.com/micmog/CIPHER
- https://mooseframework.inl.gov/modules/phase_field/
- https://openphase.rub.de
- https://www.micress.de
- maybe others?

It should be noted, that I am not an expert in phase field modeling and don't know it these softwares can be compared directly with your software.
