### 2024-11-11 0bbc6eb
Dockerfile and github workflow for Docker image build and push
