### 2024-11-11 CastillonMiguel/phasefieldx#5 @CastillonMiguel
I recently added a Dockerfile and a GitHub Action to automate the building and publishing of the Docker image. You can view the specific commit here: [0bbc6eb](https://github.com/CastillonMiguel/phasefieldx/commit/0bbc6eb2936374c29c998f9472adb943e6b6afe2). The Docker container image is now available on GitHub Packages: phasefieldx container. [phasefieldx container on GitHub Packages](https://github.com/CastillonMiguel/phasefieldx/pkgs/container/phasefieldx)

Thank you for the example reference—it was incredibly helpful! I ran into some issues installing the latest dolfinx version (0.9.0) and handling VTK file dependencies for both arm64 and amd64 architectures. Additionally, there were compatibility problems with libgl1-mesa-glx, which I resolved by substituting it with libgl1 and libglx-mesa0. As a workaround, I’ve built the image specifically for linux/amd64, and it is now working correctly.
