### 2024-11-11 1e733bb
docs: add Docker image option to installation manual
