+++ docs/source/getting-started/Installation/main.rst
@@ -39,3 +39,10 @@ Follow these steps to set up your environment:
 These steps will set up all the necessary dependencies for running the code in this repository. Make sure to activate the `phasefieldx-env` environment whenever you work with this project.
 
 For more information and additional installation options, please visit the [FEniCSx Project download page](https://fenicsproject.org/download/).
+
+We also provide a pre-built docker image with FEniCSx and phasefieldx installed. You pull this image using the command
+
+.. code-block::
+
+   docker pull ghcr.io/castillonmiguel/phasefieldx:main
+   
