### 2024-11-11 CastillonMiguel/phasefieldx#5 @finsberg
If you need vtk and pyvista support you might consider using the `lab` image from dolfinx as your base image instead (https://github.com/FEniCS/dolfinx/pkgs/container/dolfinx%2Flab) since it already comes with this pre-installed, see https://github.com/FEniCS/dolfinx/blob/70cfbeefa19a9545fbc0e37ec518cdf111e262cc/docker/Dockerfile.end-user#L152
