### 2024-11-11 CastillonMiguel/phasefieldx#6 @CastillonMiguel
Thanks for providing all these references. I would like to clarify that, at the moment, the main focus of the PhaseFieldX package is on phase field fracture and fatigue models, and the comparisons I have made in the paper are based on these aspects.

However, the package does include a simple model related to the Allen-Cahn equation, which can be compared to some of the references you provided. I can add a comment on this part in the paper.
