### 2024-11-11 CastillonMiguel/phasefieldx#5 @CastillonMiguel
I’ve added the specific parts related to PyVista and VTK to my existing Dockerfile rather than switching to the dolfinx lab image. I appreciate the tip, though. You can find the changes in my latest commit.

https://github.com/CastillonMiguel/phasefieldx/commit/1f0a7d27ccc37309fd2db7589a16247ce74888a4
