### 2024-11-12 0a2e1f2
add kernel specification
