### 2024-11-14 6308f86
Update comments and explanations in Elasticity examples
