+++ examples/Elasticity/plot_1100.py
@@ -1,8 +1,18 @@
 """
 .. _ref_1100:
 
-Displacement control
-^^^^^^^^^^^^^^^^^^^^
+Displacement Control Analysis
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+This script models a linear elastic problem following the theory detailed in (:ref:`theory_elasticity`).
+
+Model Overview
+--------------
+The model represents a square plate where:
+- The bottom edge is fully fixed, preventing both displacement and rotation.
+- The top edge can slide vertically, with a controlled vertical displacement applied.
+
+The geometry, boundary conditions, and mesh are depicted in the diagram below. The plate is discretized into quadrilateral finite elements, appropriate for 2D structural analysis.
 
 .. code-block::
 
@@ -20,6 +30,11 @@ Displacement control
    #     |
    #      ---X
    #  Z /
+
+
+Material Properties
+-------------------
+The material properties are summarized in the table below:
 
 +----+---------+--------+
 |    | VALUE   | UNITS  |
@@ -48,14 +63,21 @@ import os
 # -------------------------------
 from phasefieldx.Element.Elasticity.Input import Input
 from phasefieldx.Element.Elasticity.solver.solver import solve
-from phasefieldx.Boundary.boundary_conditions import bc_x, bc_y, bc_xy, get_ds_bound_from_marker
-from phasefieldx.Loading.loading_functions import loading_Txy
+from phasefieldx.Boundary.boundary_conditions import bc_xy, get_ds_bound_from_marker
 from phasefieldx.PostProcessing.ReferenceResult import AllResults
 
 
 ###############################################################################
-# Parameters definition
-# ---------------------
+# Define Simulation Parameters
+# ----------------------------
+# `Data` is an input object containing essential parameters for simulation setup
+# and result storage:
+# - `E`: Young's modulus, set to 210 kN/mm².
+# - `nu`: Poisson's ratio, set to 0.3.
+# - `save_solution_xdmf` and `save_solution_vtu`: Set to `False` and `True`, respectively,
+#   specifying the file format to save displacement results (.vtu here).
+# - `results_folder_name`: Name of the folder for saving results. If it exists,
+#   it will be replaced with a new empty folder.
 Data = Input(E=210.0,
              nu=0.3,
              save_solution_xdmf=False,
@@ -64,31 +86,52 @@ Data = Input(E=210.0,
 
 
 ###############################################################################
-# Mesh definition
+# Mesh Definition
 # ---------------
+# The mesh is a structured grid with quadrilateral elements:
+# - `divx`, `divy`: Number of elements along the x and y axes (10 each).
+# - `lx`, `ly`: Physical domain dimensions in x and y (1.0 units each).
+divx, divy = 10, 10
+lx, ly = 1.0, 1.0
 msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
                                     [np.array([0, 0]),
-                                     np.array([1, 1])],
-                                    [10, 10],
+                                     np.array([lx, ly])],
+                                    [divx, divy],
                                     cell_type=dolfinx.mesh.CellType.quadrilateral)
 
 
+###############################################################################
+# Boundary Identification
+# -----------------------
+# Boundary conditions are applied to specific regions of the domain:
+# - `bottom`: Identifies the $y=0$ boundary.
+# - `top`: Identifies the $y=ly$ boundary.
+# `fdim` is the dimension of boundary facets (1D for a 2D mesh).
 def bottom(x):
     return np.isclose(x[1], 0)
 
-
 def top(x):
-    return np.isclose(x[1], 1)
-
+    return np.isclose(x[1], lx)
 
[…]
