+++ examples/Elasticity/plot_1101.py
@@ -4,6 +4,16 @@
 Force control
 ^^^^^^^^^^^^^
 
+This script simulates a linear elastic problem using a force-control approach, based on the theory outlined in (:ref:`theory_elasticity`).
+
+Model Overview
+--------------
+The model represents a square plate with:
+- A fully fixed bottom edge, preventing both displacement and rotation.
+- A controlled vertical force applied along the top edge.
+
+The geometry, boundary conditions, and mesh structure are shown below. The plate is discretized using quadrilateral finite elements, suitable for 2D structural analysis.
+
 .. code-block::
 
    #           F/\/\/\/\/\/\       
@@ -21,6 +31,10 @@ Force control
    #      ---X
    #  Z /
 
+
+Material Properties
+-------------------
+The material properties are summarized in the table below:
 
 +----+---------+--------+
 |    | VALUE   | UNITS  |
@@ -50,14 +64,22 @@ import os
 # -------------------------------
 from phasefieldx.Element.Elasticity.Input import Input
 from phasefieldx.Element.Elasticity.solver.solver import solve
-from phasefieldx.Boundary.boundary_conditions import bc_x, bc_y, bc_xy, get_ds_bound_from_marker
+from phasefieldx.Boundary.boundary_conditions import bc_xy, get_ds_bound_from_marker
 from phasefieldx.Loading.loading_functions import loading_Txy
 from phasefieldx.PostProcessing.ReferenceResult import AllResults
 
 
 ###############################################################################
-# Parameters definition
-# ---------------------
+# Define Simulation Parameters
+# ----------------------------
+# `Data` is an input object containing essential parameters for simulation setup
+# and result storage:
+# - `E`: Young's modulus, set to 210 kN/mm².
+# - `nu`: Poisson's ratio, set to 0.3.
+# - `save_solution_xdmf` and `save_solution_vtu`: Set to `False` and `True`, respectively,
+#   specifying the file format to save displacement results (.vtu here).
+# - `results_folder_name`: Name of the folder for saving results. If it exists,
+#   it will be replaced with a new empty folder.
 Data = Input(E=210.0,
              nu=0.3,
              save_solution_xdmf=False,
@@ -66,31 +88,52 @@ Data = Input(E=210.0,
 
 
 ###############################################################################
-# Mesh definition
+# Mesh Definition
 # ---------------
+# The mesh is a structured grid with quadrilateral elements:
+# - `divx`, `divy`: Number of elements along the x and y axes (10 each).
+# - `lx`, `ly`: Physical domain dimensions in x and y (1.0 units each).
+divx, divy = 10, 10
+lx, ly = 1.0, 1.0
 msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
                                     [np.array([0, 0]),
-                                     np.array([1, 1])],
-                                    [10, 10],
+                                     np.array([lx, ly])],
+                                    [divx, divy],
                                     cell_type=dolfinx.mesh.CellType.quadrilateral)
 
 
+###############################################################################
+# Boundary Identification
+# -----------------------
+# Boundary conditions and forces are applied to specific regions of the domain:
+# - `bottom`: Identifies the $y=0$ boundary.
+# - `top`: Identifies the $y=ly$ boundary.
+# `fdim` is the dimension of boundary facets (1D for a 2D mesh).
 def bottom(x):
     return np.isclose(x[1], 0)
 
-
 def top(x):
-    return np.isclose(x[1], 1)
-
+    return np.isclose(x[1], lx)
 
 fdim = msh.topology.dim - 1
 
+# Using the `bottom` and `top` functions, we locate the facets on the bottom and top sides of the mesh,
+# where $y = 0$ and $y = ly$, respectively. The `locate_entities_boundary` function returns an array of facet
[…]
