### 2024-11-14 823c48c
Update comments and explanations in Allen Cahn example
