+++ examples/AllenCahn/plot_4000.py
@@ -4,7 +4,7 @@ r"""
 Free Energy: Allen-Cahn
 ^^^^^^^^^^^^^^^^^^^^^^^
 
-In this example, we study the free energy potential of the Allen-Cahn equation. The model consists of a bar of length $2a$ with boundary conditions at $x=-a$ of $\phi=-1$ and $\phi=1$ at $x=a$.
+In this example, we study the free energy potential of the Allen-Cahn equation. The model consists of a bar of length $2a$ with boundary conditions at $x=-a$ of $\phi=-1$ and $\phi=1$ at $x=a$. The theoric framework is explained in (:ref:`theory_allen_cahn`).
 
 .. code-block::
 
@@ -72,8 +72,21 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 ###############################################################################
 # Parameters definition
 # ---------------------
-# A length scale parameter of $l = 10.0$ is defined. The phase field is saved in a VTU file.
-# All the results are stored in a folder named results_folder_name = name.
+# First, we define an input class, which contains all the parameters needed for the setup 
+# and results of the simulation.
+#
+# The first term, $l$, specifies the length scale parameter for the problem, with $l = 10.0$.
+#
+# The next two options, `save_solution_xdmf` and `save_solution_vtu`, determine the file format
+# used to save the phase-field results (.xdmf or .vtu), which can then be visualized using
+# tools like ParaView or pvista. Both parameters are boolean values (`True` or `False`).
+# In this case, we set `save_solution_vtu=True` to save the phase-field results in .vtu format.
+#
+# Lastly, `results_folder_name` specifies the name of the folder where all results
+# and log information will be saved. If the folder does not exist, `phasefieldx` will create it. 
+# However, if the folder already exists, any previous data in it will be removed, and a new 
+# empty folder will be created in its place.
+#
 Data = Input(l=10.0,
              save_solution_xdmf=False,
              save_solution_vtu=True,
@@ -84,9 +97,12 @@ Data = Input(l=10.0,
 # Mesh Definition
 # ---------------
 # A 2a mm x 1 mm rectangular mesh with quadrilateral elements is created using Dolfinx.
+# The mesh is a structured grid with quadrilateral elements:
+# - `divx`, `divy`: Number of elements along the x and y axes .
+# - `lx`, `ly`: Physical domain dimensions in x and y.
 a = 50.0
-divx, divy, divz = 100, 1, 1
-lx, ly, lz = 100.0, 1.0, 1.0
+divx, divy = 100, 1
+lx, ly = 100.0, 1.0
 
 msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
                                     [np.array([-a, -0.5]),
@@ -96,22 +112,37 @@ msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
 
 
 ###############################################################################
-# The left and right parts are identified and used to impose
-# the boundary conditions.
+# Boundary Identification
+# -----------------------
+# Boundary conditions and forces are applied to specific regions of the domain:
+# - `left`: Identifies the $y=-a$ boundary.
+# - `right`: Identifies the $y=a$ boundary.
+# `fdim` is the dimension of boundary facets (1D for a 2D mesh).
 def left(x):
     return np.isclose(x[0], -a)
-
 
 def right(x):
     return np.isclose(x[0], a)
 
-
 fdim = msh.topology.dim - 1
 
+# Using the `bottom` and `top` functions, we locate the facets on the left and right sides of the mesh,
+# where $x = -a$ and $x = a$, respectively. The `locate_entities_boundary` function returns an array of facet
+# indices representing these identified boundary entities.
 left_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, left)
[…]
