+++ joss/paper.md
@@ -51,7 +51,7 @@ Note that multiple types of fatigue degradation functions have been developed [@
 
 One of the key aspects of phase-field fracture simulations is that the crack surface density functional provides a continuous approximation of cracks, which converges to a discrete crack representation as the length scale parameter approaches zero. This corresponds to the second part of the phase-field fracture functional, as given in \autoref{eq:phase_field_fracture_functional}. Therefore, before studying phase-field fracture models, it is recommended to first focus on this part of the functional in isolation.
 
-The second part of the functional, which controls the regularization of the phase-field, is expressed as:
+The second part of the functional, which governs the regularization of the phase-field, is presented in \autoref{eq:crack_representation_functional}.
 \begin{equation}\label{eq:crack_representation_functional}
 W[\phi] = \int_\Omega \left( \frac{1}{2l} \phi^2 + \frac{l}{2} |\nabla \phi|^2 \right) \mathrm{d}\Omega
 \end{equation}
@@ -62,7 +62,7 @@ By considering this, an analytical solution can be derived for a one-dimensional
 \end{equation}
 where $l$ is the length scale parameter and $a$ is half the length of the bar. In the limiting case where $l \to 0$, it follows that $a/l \to \infty$, and the phase-field is represented solely by the $e^{-|x|/l}$ term. The total energy of the bar is given by $W=\tanh(a/l)$, and in the limit case as $l \to 0$, $W$ approaches $1$.
 
-Note that as the length scale parameter tends to zero, the continuous phase-field representation converges to a discrete crack representation. \autoref{fig:bar_one_dimension_solution} illustrates the crack approximation for two different length scale values.
+Note that as the length scale parameter tends to zero, the continuous phase-field representation converges to a discrete crack representation. \autoref{fig:bar_one_dimension_solution} illustrates the crack approximation for a bar of length $2a$ (with $a=1$), for two different length-scale values.
 
 ![Left: Cracked bar represented by the phase-field variable. Right: Phase-field representation of the cracked bar with $l=0.1$ and $l=0.5$.\label{fig:bar_one_dimension_solution}](./images/crack_surface_1d_solution.png){width="360pt"}
 
@@ -72,9 +72,13 @@ Although the main capabilities of **PhaseFieldX** focus on fracture, fatigue, cr
 
 # Statement of Need
 
-**PhaseFieldX** stands out in the landscape of phase-field modeling software as an open-source package, with no commercial solutions currently available that directly incorporate phase-field fracture simulations; the primary alternative is the use of *ABAQUS* subroutines, which necessitate custom coding by the user. Most other software options are in-house codes that can be difficult to share, replicate, or adapt. While some implementations are present in open-source software like *FEniCS* or *Comsol*, these are often limited to specific problems and specific implementations of the phase-field model. In contrast, **PhaseFieldX** integrates the latest advancements in phase-field research, including fatigue simulations, into a user-friendly, ready-to-use package for scientific research. This accessibility encourages broader adoption of phase-field methods, simplifying the process for researchers to engage in cutting-edge investigations without the complexities of bespoke modeling tools.
+**PhaseFieldX** distinguishes itself as a leading open-source package for phase-field modeling, particularly in the absence of commercial solutions that directly support phase-field fracture simulations. The primary alternatives include the use of *ABAQUS* subroutines, which require users to develop custom code, or open-source codes such as *RACCOON* [@RACCOON], a parallel finite-element framework specialized in phase-field fracture simulations. 
 
[…]
