+++ joss/paper.bib
@@ -41,70 +41,121 @@
 
 % Phase-field fracture / fatigue ======================================================
 @article{Miehe2010,
-title = {A phase field model for rate-independent crack propagation: Robust algorithmic implementation based on operator splits},
-journal = {Computer Methods in Applied Mechanics and Engineering},
-volume = {199},
-number = {45},
-pages = {2765-2778},
-year = {2010},
-issn = {0045-7825},
-doi = {10.1016/j.cma.2010.04.011},
-url = {https://www.sciencedirect.com/science/article/pii/S0045782510001283},
-author    = {Miehe, C. and Hofacker, M. and Welschinger, F.},
+  title     = {A phase field model for rate-independent crack propagation: Robust algorithmic implementation based on operator splits},
+  author    = {Miehe, C. and Hofacker, M. and Welschinger, F.},
+  journal   = {Computer Methods in Applied Mechanics and Engineering},
+  volume    = {199},
+  number    = {45},
+  pages     = {2765-2778},
+  year      = {2010},
+  issn      = {0045-7825},
+  doi       = {10.1016/j.cma.2010.04.011},
+  url       = {https://www.sciencedirect.com/science/article/pii/S0045782510001283},
 }
 
 @article{FrancfortMarigo1998,
-title = {Revisiting brittle fracture as an energy minimization problem},
-journal = {Journal of the Mechanics and Physics of Solids},
-volume = {46},
-number = {8},
-pages = {1319-1342},
-year = {1998},
-issn = {0022-5096},
-doi = {10.1016/S0022-5096(98)00034-9},
-url = {https://www.sciencedirect.com/science/article/pii/S0022509698000349},
-author = {Francfort, G.A. and Marigo, J.-J.},
+  title     = {Revisiting brittle fracture as an energy minimization problem},
+  author    = {Francfort, G.A. and Marigo, J.-J.},
+  journal   = {Journal of the Mechanics and Physics of Solids},
+  volume    = {46},
+  number    = {8},
+  pages     = {1319-1342},
+  year      = {1998},
+  issn      = {0022-5096},
+  doi       = {10.1016/S0022-5096(98)00034-9},
+  url       = {https://www.sciencedirect.com/science/article/pii/S0022509698000349},
 }
 
 @article{Amor2009,
-title = {Regularized formulation of the variational brittle fracture with unilateral contact: Numerical experiments},
-journal = {Journal of the Mechanics and Physics of Solids},
-volume = {57},
-number = {8},
-pages = {1209-1229},
-year = {2009},
-issn = {0022-5096},
-doi = {10.1016/j.jmps.2009.04.011},
-url = {https://www.sciencedirect.com/science/article/pii/S0022509609000659},
-author    = {Amor, H. and Marigo, J.-J. and Maurini, C.},
-keywords = {Fracture, Energy methods, Free-discontinuity problems, Contact mechanics, Finite elements}
+  title     = {Regularized formulation of the variational brittle fracture with unilateral contact: Numerical experiments},
+  author    = {Amor, H. and Marigo, J.-J. and Maurini, C.},
+  journal   = {Journal of the Mechanics and Physics of Solids},
+  volume    = {57},
+  number    = {8},
+  pages     = {1209-1229},
+  year      = {2009},
+  issn      = {0022-5096},
+  doi       = {10.1016/j.jmps.2009.04.011},
+  url       = {https://www.sciencedirect.com/science/article/pii/S0022509609000659},
+  keywords  = {Fracture, Energy methods, Free-discontinuity problems, Contact mechanics, Finite elements}
 }
 
-
 @Article{Ambati2015,
[…]
