### 2024-11-15 CastillonMiguel/phasefieldx#6 @CastillonMiguel
I added some of the references you suggested, and I also added some related to phase-field fracture simulation (the Raccoon GitHub repository https://github.com/hugary1995/raccoon ). I also made some minor changes to the paper. I updated the JOSS paper in the commit https://github.com/CastillonMiguel/phasefieldx/commit/0ecf1f6d128da74a8c06109278a064499e1dd13b, and the paper was also generated by GitHub Actions in: https://github.com/CastillonMiguel/phasefieldx/actions/workflows/build_joss_paper.yml
