### 2024-11-15 CastillonMiguel/phasefieldx#9 @CastillonMiguel
I’m currently working on revising the examples and adding more detailed text, including necessary references and a clearer explanation of each example's purpose.
To keep track of the progress, I’m planning to address each examples section , and here’s a list of sections that I’ve revised or am working on revising:

* ~~AllenCahn:~~ https://github.com/CastillonMiguel/phasefieldx/commit/823c48c481359e67fab5b68e4f811855d2619005
* ~~Elasticity:~~ https://github.com/CastillonMiguel/phasefieldx/commit/6308f86429a04a94abc898a2a93a1f3fe1cb5f43
* Fatigue
* GmshGeoFiles
* ~~PhaseField:~~ https://github.com/CastillonMiguel/phasefieldx/commit/e79cd38ed7d23015ef0629e1300b06558a21d880
* PhaseFieldFracture

I’ll update the examples to ensure they form coherent, self-contained explanations.
Thanks again for your input!
