### 2024-11-15 e870130
Update comments and explanations in GmshGeoFiles examples
