+++ examples/GmshGeoFiles/plot_9101.py
@@ -3,6 +3,13 @@
 
 .geo File: Single Notched Tension Test
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+This example demonstrates the procedure for generating a mesh for a single edge notched tension test.
+Note that Phasefieldx can import external meshes in the .msh format. This can be achieved by using Gmsh.
+
+Files with the .geo format define both the geometry and the mesh parameters, such as element types and sizes.
+Gmsh can then generate the mesh based on the input from these .geo files.
+
+Below is an example of the .geo file used for mesh generation.
 
 .geo file
 ---------
@@ -10,6 +17,16 @@
 .. include::  ../../../../examples/GmshGeoFiles/9101_SingleNotchedTensionTest/file.geo
    :literal:
 
+For more information about Gmsh and the `.geo` file format, please refer to the official `Gmsh website <https://gmsh.info>`_.
+
+To generate the mesh in the `.msh` format from the terminal, use the following Gmsh command:
+
+.. code-block::
+   gmsh file.geo -3 -o mesh.msh
+
+Here, `-3` specifies that the mesh is 3-dimensional, and `-o mesh.msh` tells Gmsh to save the mesh to a file named `mesh.msh` with the `.msh` extension.
+
+Alternatively, you can generate the mesh using the Gmsh Python API, which allows for programmatic mesh generation within Python scripts.
 """
 
 ###############################################################################
@@ -34,7 +51,10 @@ gmsh.open(geo_file)
 # Generate the mesh (2D example, for 3D use generate(3))
 gmsh.model.mesh.generate(3)
 
-# Write the mesh to a .vtu file
+# Write the mesh to a .vtk file for visualization
+# Note that the input mesh file for the *phasefieldx* simulation should have the .msh extension.
+# Use "output_mesh_for_view.msh" to generate the mesh for the simulation input.
+# In this case, the mesh is saved in .vtk format to facilitate visualization with PyVista.
 vtu_file = os.path.join(folder, "output_mesh_for_view.vtk")
 gmsh.write(vtu_file)
 
