### 2024-11-20 c6f8b60
Update comments and explanations in Fatigue example
