+++ examples/Fatigue/plot_1800.py
@@ -4,19 +4,19 @@ r"""
 Fatigue: Single edge notched tension test
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-A well-known benchmark simulation in fracture mechanics is performed, relying on the simulation conducted by [Carrara]_. This simulation considers an isotropic formulation.
+In this study, a phase-field fatigue simulation is analyzed. The theoretical foundation of this model is presented in (:ref:theory_phase_field_fracture). Specifically, a cyclic displacement is applied to a single-edge notched tension test, following the approach described by [Carrara]_. The simulation adopts an isotropic formulation.
 
-The model consists of a square plate with a notch located halfway up, extending from the left to the center, as shown in the figure below. The bottom part is fixed in all directions, while the upper part can slide vertically. A vertical displacement is applied at the top. The geometry and boundary conditions are depicted in the figure. We discretize the model with triangular elements, refining the areas (element size h) where crack evolution is expected. The element size h must be sufficiently small to avoid mesh dependencies.
+The model consists of a square plate with a notch located midway along the left edge, extending horizontally toward the center, as illustrated in the figure below. The bottom edge of the plate is fixed in all directions, while the top edge is free to slide vertically. A cyclic vertical displacement is applied to the top edge. The geometry and boundary conditions are clearly shown in the accompanying figure. The model is discretized using triangular finite elements, with refined mesh resolution (element size $h$) in regions where crack evolution is anticipated. The element size $h$ must be sufficiently small to minimize mesh dependency.
 
-A cyclic tensile test is conducted. A symmetric cyclic load is applied with a displacement amplitude of $\Delta u = 4 \times 10^{-3} mm$. The results are presented in terms of the accumulation of the fatigue history variable $\bar{\alpha}$ versus the number of cycles N (fatigue life curves).
+A cyclic tensile test is performed under symmetric cyclic loading, with a displacement amplitude of $\Delta u = 4 \times 10^{-3} mm$. The results are presented as fatigue life curves, showing the accumulation of the fatigue history variable $\bar{\alpha}$ as a function of the number of cycles $N$.
 
 .. code-block::
 
-   #           u/\/\/\/\/\/\
-   #            ||||||||||||
-   #            *----------*
-   #            |          |
-   #            | a=0.5    |
+   #                           /\
+   #                           ||
+   #            *----------*  ---- u
+   #            |          |   ||
+   #            | a=0.5    |   \/
    #            |---       |
    #            |          |
    #            |          |
@@ -28,6 +28,9 @@ A cyclic tensile test is conducted. A symmetric cyclic load is applied with a di
    #  Z /
 
 
+Material Properties
+-------------------
+The material properties, length scale parameter, and fatigue parameter are summarized in the table below:
 
 +----------+---------+--------+
 |          | VALUE   | UNITS  |
@@ -72,12 +75,29 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 
 
 ###############################################################################
-# Parameters Definition
-# ---------------------
-# Here, we define the class containing the input parameters. The material parameters are:
-# Young's modulus $E = 210$, $\text{kN/mm}^2$, Poisson's ratio $\nu = 0.3$,
[…]
