### 2024-11-21 be53181
Add # %% markers in Python examples for improved DOCS view
