+++ examples/GmshGeoFiles/plot_9104.py
@@ -42,16 +42,21 @@ import gmsh
 import pyvista as pv
 
 folder = "9104_Structured"
+
+# %%
 # Initialize Gmsh
 gmsh.initialize()
 
+# %%
 # Open the .geo file
 geo_file = os.path.join(folder, "file.geo")
 gmsh.open(geo_file)
 
+# %%
 # Generate the mesh (2D example, for 3D use generate(3))
 gmsh.model.mesh.generate(2)
 
+# %%
 # Write the mesh to a .vtk file for visualization
 # Note that the input mesh file for the *phasefieldx* simulation should have the .msh extension.
 # Use "output_mesh_for_view.msh" to generate the mesh for the simulation input.
@@ -59,6 +64,7 @@ gmsh.model.mesh.generate(2)
 vtu_file = os.path.join(folder, "output_mesh_for_view.vtk")
 gmsh.write(vtu_file)
 
+# %%
 # Finalize Gmsh
 gmsh.finalize()
 
