### 2024-11-21 a1deeb6
Update comments and explanations in PhaseFieldFracture examples
