+++ examples/PhaseFieldFracture/plot_1700.py
@@ -4,6 +4,8 @@ r"""
 One Element tension Isotropic
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+In this example, we consider a phase-field fracture simulation. The theoretical basis for this model is outlined in detail in (:ref:`theory_phase_field_fracture`).  For this simulation, we do not consider split energy; instead, the isotropic model is used. The setup involves a simple model comprising a single four-node element with dimensions of 1x1 mm. The bottom nodes are fully constrained in both directions, while the top nodes are allowed to slide vertically.
+
 .. code-block::
 
    #          u /\        /\
@@ -21,6 +23,8 @@ One Element tension Isotropic
    #     *---X
 
 
+The Young's modulus, Poisson's ratio, and the critical energy release rate are given in the table :ref:`Properties <table_properties_label>`. Young's modulus $E$ and Poisson's ratio $\nu$ can be represented with the Lamé parameters as: $\lambda=\frac{E\nu}{(1+\nu)(1-2\nu)}$; $\mu=\frac{E}{2(1+\nu)}$.
+
 .. _table_properties_label:
 
 .. table:: Properties
@@ -37,18 +41,18 @@ One Element tension Isotropic
    | l  | 0.1     | mm     |
    +----+---------+--------+
 
-In this example, we study a simple model formed by a single element with four nodes and dimensions of 1x1 mm. The bottom nodes are constrained in both directions, and the top nodes can slide vertically.
-
-The Young's modulus, Poisson's ratio, and the critical energy release rate are given in the table :ref:`Properties <table_properties_label>`. Young's modulus $E$ and Poisson's ratio $\nu$ can be represented with the Lamé parameters as: $\lambda=\frac{E\nu}{(1+\nu)(1-2\nu)}$; $\mu=\frac{E}{2(1+\nu)}$.
-
 In this case, due to the discretization, it is possible to obtain an analytical solution for the isotropic model by solving $\phi$ from the given equations. The term $|\nabla \phi|^2$ vanishes due to the discretization as explained by Molnar \cite{MOLNAR201727} and Miehe \cite{Miehe1} in the appendix.
 
 .. math::
    \phi = \frac{2 \psi_a}{\frac{G_c}{l}+2\psi_a}=\frac{2 H}{\frac{G_c}{l}+2H}
 
 .. math::
-    \sigma_y = \sigma_{a}(1-\phi)^2 + \sigma_{b}
+    \sigma_y = \sigma(1-\phi)^2
 
+
+.. [MOLNAR201727] 2D and 3D Abaqus implementation of a robust staggered phase-field solution for modeling brittle fracture,  https://doi.org/10.1016/j.finel.2017.03.002
+
+.. [Miehe1] A phase field model for rate-independent crack propagation: Robust algorithmic implementation based on operator splits, https://doi.org/10.1016/j.cma.2010.04.011.
 """
 
 ###############################################################################
@@ -75,10 +79,27 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 ###############################################################################
 # Parameters Definition
 # ---------------------
-# Here, we define the class containing the input parameters. The material parameters are:
-# Young's modulus $E = 210$, $\text{kN/mm}^2$, Poisson's ratio $\nu = 0.3$,
-# critical energy release rate $G_c = 0.005$, $\text{kN/mm}^2$, and length scale parameter $l = 0.1$, $\text{mm}$.
-# We consider isotropic degradation where all the energy is degraded and irreversibility as proposed by Miehe.
+# `Data` is an input object containing essential parameters for simulation setup
+# and result storage:
[…]
