+++ examples/PhaseFieldFracture/plot_1712.py
@@ -6,7 +6,7 @@ Single edge notched shear test
 
 A well-known benchmark simulation in fracture mechanics is performed, based on the work by [Miehe]_.
 
-The model consists of a square plate with a notch located halfway up, extending from the left to the center, as shown in the figure below. The bottom part is fixed in all directions, while the upper part can slide vertically. A horizontal displacement is applied at the top. The geometry and boundary conditions are depicted in the figure. We discretize the model by refining the areas where crack evolution is expected, ensuring the element size h is sufficiently small to avoid mesh dependencies.
+The model consists of a square plate with a notch located halfway up, extending from the left side to the center, as shown in the figure below. The bottom part is fixed in all directions, while the top part can slide horizontally. A horizontal displacement is applied at the top. The geometry and boundary conditions are depicted in the figure. We discretize the model by refining the areas where crack evolution is expected, ensuring that the element size $h$ is sufficiently small to avoid mesh dependency.
 
 .. code-block::
 
@@ -25,6 +25,10 @@ The model consists of a square plate with a notch located halfway up, extending 
    #      ---X
    #  Z /
 
+
+The Young's modulus, Poisson's ratio, and the critical energy release rate are given in the table :ref:`Properties <table_properties_label>`. Young's modulus $E$ and Poisson's ratio $\nu$ can be represented with the Lamé parameters as: $\lambda=\frac{E\nu}{(1+\nu)(1-2\nu)}$; $\mu=\frac{E}{2(1+\nu)}$.
+
+.. _table_properties_label:
 
 +----+---------+--------+
 |    | VALUE   | UNITS  |
@@ -66,46 +70,86 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 ###############################################################################
 # Parameters Definition
 # ---------------------
-# Here, we define the class containing the input parameters. The material parameters are:
-# Young's modulus $E = 210$, $\text{kN/mm}^2$, Poisson's ratio $\nu = 0.3$,
-# critical energy release rate $G_c = 0.0027$, $\text{kN/mm}^2$, and length scale parameter $l = 0.015$, $\text{mm}$.
-# We consider anisotropic degradation (spectral) and irreversibility as proposed by Miehe.
+# `Data` is an input object containing essential parameters for simulation setup
+# and result storage:
+# - `E`: Young's modulus, set to 210 $kN/mm^2$.
+# - `nu`: Poisson's ratio, set to 0.3.
+# - `Gc`: Critical energy release rate, set to 0.005 $kN/mm$.
+# - `l`: Length scale parameter, set to 0.1 $mm$.
+# - `degradation`: Specifies the degradation type. Options are "isotropic" or "anisotropic".
+# - `split_energy`: Controls how the energy is split; options include "no" (default), "spectral," or "deviatoric." 
+# In this case an anisotropic formulation with volumetric-deviatric decomposition is considered.
+# - `degradation_function`: Specifies the degradation function; here, it is "quadratic."
+# - `irreversibility`: Determines the irreversibility criterion; in this case, set to "miehe."
+# - `fatigue`: Enables fatigue simulation when set to `True`.
+# - `fatigue_degradation_function`: Defines the function for fatigue degradation, set to "asymptotic."
+# - `fatigue_val`: Fatigue parameter value (used only in fatigue simulations, not in this one).
+# - `k`: Stiffness penalty parameter, set to 0.0.
+# - `min_stagger_iter`: Minimum number of staggered iterations, set to 2.
+# - `max_stagger_iter`: Maximum number of staggered iterations, set to 500.
+# - `stagger_error_tol`: Error tolerance for staggered iterations, set to 1e-8.
[…]

[…]
