### 2024-11-21 73261ab
add pv.start_xvfb() command
