+++ examples/PhaseFieldFracture/plot_1712.py
@@ -317,6 +317,7 @@ S.set_color('b')
 # The phase-field result saved in the .vtu file is shown.
 # For this, the file is loaded using PyVista.
 file_vtu = pv.read(os.path.join(Data.results_folder_name, "paraview-solutions_vtu", "phasefieldx_p0_000097.vtu"))
+pv.start_xvfb()
 file_vtu.plot(scalars='phi', cpos='xy', show_scalar_bar=True, show_edges=False)
 
 
