### 2024-11-28 c395475
Refine example explanations for clarity
