### 2024-11-28 dcc9de2
Add example mesh for the Damage Mechanics Challenge
