### 2024-12-02 f385f6a
Include new challenge test example
