+++ examples/PhaseFieldFracture/plot_1718.py
@@ -1,0 +1,435 @@
+"""
+.. _ref_1718:
+
+Damage Mechanics Challenge
+^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+This example presents a three-dimensional simulation based on the *Purdue-SANDIA-LLNL Damage Mechanics Challenge*. For more details, visit the `Damage Mechanics Challenge Webpage <https://purr.purdue.edu/groups/damagemechanicschallenge>`_. The geometry, parameters, and considerations used in this example are derived from the challenge specifications.
+
+.. image:: ../../../../examples/PhaseFieldFracture/1718_Damage_Challenge/animation.gif
+
+Several studies related to this challenge and phase-field fracture simulations are available in the literature, including [Yousef]_, [Abhinav]_, and [Jonas]_.
+
+Although the primary objective of this challenge is to predict damage evolution and failure of the specimen, the focus of this example is to demonstrate a three-dimensional simulation using the phase-field fracture model.
+
+A rectangular plate with an initial notch positioned at the bottom center is used in this example, as shown in the figure below. This beam is supported at its ends, with the bottom-left part fixed in all directions and the bottom-right part fixed in the vertical direction. A vertical displacement is applied at the top. The geometry and boundary conditions are depicted in the figure.
+
+The model is discretized using tetrahedral elements, with refined mesh sizes (denoted as `h`) in regions where crack evolution is expected. The element size `h` must be sufficiently small to prevent mesh dependencies and ensure accurate results.
+
+.. code-block::
+
+   #                      <------------------------- 76.2--------------------------> 
+   #                                                                  <--- 24.96---> 
+   #                 /\   *---------------------||||-------------------------------*  /\
+   #                 /   /                 u    \/\/            _____            / |  |
+   #        12.7    /   /                     ||||             /<1> /|          /  |  | 
+   #               /   /                      \/\/            /    / |         /   |  | 25.4
+   #              \/  *-------------------------------------/-----/--|--------*    |  |
+   #                  |                                    /   /   /          |    |  \/ 
+   #                  |                                  /   /   /            |    *  
+   #                  |                                 /  /   /              |   /   
+   #                  |                                /__/  /                |  /   
+   #                  |                                | | /                  | /   
+   #                  *--------------------------------. .--------------------*  
+   #                         /_\/_\                    <1>             /_\/_\       
+   #                        ///////                                  oo  oo
+   #                  <--7.62-->                                    <--7.62-->
+   #                  <------------- 43.89 ----------> 
+   #    |Y            
+   #    ---X          
+   # Z /         
+
+.. note::
+    The mesh can be generated using Gmsh, as demonstrated in the example :ref:`ref_9107`.
+   
+The Young's modulus, Poisson's ratio, and the critical energy release rate are given in the table :ref:`Properties <table_properties_label>`. Young's modulus $E$ and Poisson's ratio $\nu$ can be represented with the Lamé parameters as: $\lambda=\frac{E\nu}{(1+\nu)(1-2\nu)}$; $\mu=\frac{E}{2(1+\nu)}$.
+
+.. _table_properties_label:
+
++----+---------+--------+
+|    | VALUE   | UNITS  |
++====+=========+========+
+| E  | 0.6     | kN/mm2 |
++----+---------+--------+
+| nu | 0.2     | [-]    |
++----+---------+--------+
[…]
