### 2024-12-02 af02376
update pyvista plots
