### 2024-12-02 003cade
Correct punctuation and spacing in the model description
