+++ examples/PhaseFieldFracture/plot_1718.py
@@ -20,7 +20,7 @@ The model is discretized using tetrahedral elements, with refined mesh sizes (de
 
    #                      <------------------------- 76.2--------------------------> 
    #                                                                  <--- 24.96---> 
-   #                 /\   *---------------------||||-------------------------------*  /\
+   #                 /\   *---------------------||||-------------------------------*  /\ 
    #                 /   /                 u    \/\/            _____            / |  |
    #        12.7    /   /                     ||||             /<1> /|          /  |  | 
    #               /   /                      \/\/            /    / |         /   |  | 25.4
