+++ docs/source/api/Logger/main.rst
@@ -1,0 +1,7 @@
+Logger
+======
+
+.. automodule:: phasefieldx.Logger.library_versions
+   :members:
+   :undoc-members:
+   :show-inheritance:
