### 2024-12-20 CastillonMiguel/phasefieldx#9 @finsberg
Looks very good! A few minor things
- The plot at the end [here](https://phasefieldx.readthedocs.io/en/latest/auto_examples/PhaseField/plot_2000.html) is not showing.
- Some text on [this page](https://phasefieldx.readthedocs.io/en/latest/auto_examples/PhaseFieldFracture/plot_1714.html#boundary-conditions) looks now well formatted.
