### 2024-12-20 #7307 @finsberg
@zhubonan I have completed my review and can recommend PhaseFieldX for publication in JOSS. PhaseFieldX is very well documented software with a lot to good material to learn about phasefield modeling. The author has been very responsive and addressed the issues being raised. I would, however encourage the author to run a linter / formatter as part of the continous integration to be consistent about formatting and to remove unused code (such as unused imports). Also, I would encourage the author to use pull requests rather than committing and pushing directly to the main branch. However, in my opinion, these issues (as well as https://github.com/CastillonMiguel/phasefieldx/issues/2) should not be blocking it from acceptance in JOSS.
