### 2024-12-21 67f3a5e
Remove LaTeX rendering configuration for plot labels
