### 2024-12-21 3023718
docs: add raw string prefix to docstring
