+++ examples/PhaseField/plot_2001.py
@@ -1,4 +1,4 @@
-"""
+r"""
 .. _ref_2001:
 
 Representation of a Cracked Plate
