### 2024-12-21 3acc562
minor change: format legend
