### 2024-12-21 1386e33
uncomment pv.start_xvfb()
