### 2024-12-21 7c81198
minor change: set \nabla as V symbol
