### 2025-01-09 0b1fb87
Fix formatting issues in documentation and update example scripts to improve consistent reaction force calculations
