+++ examples/PhaseFieldFracture/plot_1713.py
@@ -1,4 +1,4 @@
-"""
+r"""
 .. _ref_1713:
 
 Symmetry: Single edge notched tension test
@@ -70,6 +70,7 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 # ---------------------
 # `Data` is an input object containing essential parameters for simulation setup
 # and result storage:
+#
 # - `E`: Young's modulus, set to 210 $kN/mm^2$.
 # - `nu`: Poisson's ratio, set to 0.3.
 # - `Gc`: Critical energy release rate, set to 0.005 $kN/mm$.
@@ -102,7 +103,7 @@ Data = Input(E=210.0,   # young modulus
              fatigue_val=0.0,
              k=0.0,
              min_stagger_iter=2,
-             max_stagger_iter=600,
+             max_stagger_iter=5000,
              stagger_error_tol=1e-8,
              save_solution_xdmf=False,
              save_solution_vtu=True,
@@ -113,9 +114,10 @@ Data = Input(E=210.0,   # young modulus
 # Mesh Definition
 # ---------------
 # The mesh is a structured grid with quadrilateral elements:
+#
 # - `divx`, `divy`: Number of elements along the x and y axes.
 # - `lx`, `ly`: Physical domain dimensions in x and y.
-divx, divy = 160, 80
+divx, divy = 200, 100
 lx, ly = 1.0, 0.5
 
 msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
@@ -129,9 +131,10 @@ msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
 # Boundary Identification
 # -----------------------
 # Boundary conditions are applied to specific regions of the domain:
+#
 # - `bottom`: Identifies the $y=0$ boundary.
 # - `top`: Identifies the $y=ly$ boundary.
-# `fdim` is the dimension of boundary facets (1D for a 2D mesh).
+# - `fdim` is the dimension of boundary facets (1D for a 2D mesh).
 def bottom(x):
     return np.logical_and(np.isclose(x[1], 0), np.greater(x[0], 0.5))
 
@@ -142,6 +145,7 @@ fdim = msh.topology.dim - 1 # Dimension of the mesh facets
 # %%
 # Facets defined in the .geo file used to generate the 'mesh.msh' file are identified here.
 # Each marker variable corresponds to a specific region on the specimen:
+#
 # - `bottom_facet_marker`: Refers to the bottom part of the specimen.
 # - `top_facet_marker`: Refers to the top part of the specimen.
 bottom_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, bottom)
@@ -150,6 +154,7 @@ top_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, top)
 # %%
 # The `get_ds_bound_from_marker` function creates measures for applying boundary conditions
 # on specific facets. These measures are generated for:
+#
 # - `bottom_facet_marker` → Stored in `ds_bottom`
 # - `top_facet_marker` → Stored in `ds_top`
 ds_bottom = get_ds_bound_from_marker(bottom_facet_marker, msh, fdim)
@@ -158,15 +163,18 @@ ds_top = get_ds_bound_from_marker(top_facet_marker, msh, fdim)
 # %%
 # `ds_list` is an array that organizes boundary condition measures alongside descriptive names.
 # Each entry in `ds_list` consists of two elements:
+#
 # - A measure (e.g., `ds_bottom`)
 # - A corresponding name (e.g., `"bottom"`)
+#
 # This structure simplifies the process of saving results by associating each boundary condition
[…]
