+++ examples/PhaseFieldFracture/plot_1715.py
@@ -1,4 +1,4 @@
-"""
+r"""
 .. _ref_1715:
 
 Symmetry: Center notched tension test
@@ -70,6 +70,7 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 # ---------------------
 # `Data` is an input object containing essential parameters for simulation setup
 # and result storage:
+#
 # - `E`: Young's modulus, set to 210 $kN/mm^2$.
 # - `nu`: Poisson's ratio, set to 0.3.
 # - `Gc`: Critical energy release rate, set to 0.005 $kN/mm$.
@@ -113,6 +114,7 @@ Data = Input(E=210.0,   # young modulus
 # Mesh Definition
 # ---------------
 # The mesh is a structured grid with quadrilateral elements:
+#
 # - `divx`, `divy`: Number of elements along the x and y axes.
 # - `lx`, `ly`: Physical domain dimensions in x and y.
 divx, divy = 100, 300
@@ -123,13 +125,16 @@ msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
                                     [divx, divy],
                                     cell_type=dolfinx.mesh.CellType.quadrilateral)
 
+
+###############################################################################
 # Boundary Identification
 # -----------------------
 # Boundary conditions are applied to specific regions of the domain:
+#
 # - `bottom`: Identifies the $y=0$ boundary.
 # - `top`: Identifies the $y=3.0$ boundary.
 # - `left`: Identifies the $x=0$ boundary.
-# `fdim` is the dimension of boundary facets (1D for a 2D mesh).
+# - `fdim` is the dimension of boundary facets (1D for a 2D mesh).
 def bottom(x):
     return np.logical_and(np.isclose(x[1], 0), np.greater(x[0], 0.5))
 
@@ -146,6 +151,7 @@ fdim = msh.topology.dim - 1 # Dimension of the mesh facets
 # %%
 # Facets defined in the .geo file used to generate the 'mesh.msh' file are identified here.
 # Each marker variable corresponds to a specific region on the specimen:
+#
 # - `bottom_facet_marker`: Refers to the bottom part of the specimen.
 # - `top_facet_marker`: Refers to the top part of the specimen.
 # - `left_facet_marker`: Refers to the left part of the specimen.
@@ -156,6 +162,7 @@ left_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, left)
 # %%
 # The `get_ds_bound_from_marker` function creates measures for applying boundary conditions
 # on specific facets. These measures are generated for:
+#
 # - `bottom_facet_marker` → Stored in `ds_bottom`
 # - `top_facet_marker` → Stored in `ds_top`
 # - `left_facet_marker` → Stored in `ds_left`
@@ -166,17 +173,19 @@ ds_top = get_ds_bound_from_marker(top_facet_marker, msh, fdim)
 # %%
 # `ds_list` is an array that organizes boundary condition measures alongside descriptive names.
 # Each entry in `ds_list` consists of two elements:
+#
 # - A measure (e.g., `ds_bottom`)
 # - A corresponding name (e.g., `"bottom"`)
 # This structure simplifies the process of saving results by associating each boundary condition
 # measure with a clear label. For instance:
+#
 # - `ds_bottom` is labeled as `"bottom"`.
 # - `ds_left` is labeled as `"left"`.
 # - `ds_top` is labeled as `"top"`.
 ds_list = np.array([
+                    [ds_top, "top"],
                    [ds_bottom, "bottom"],
                    [ds_left, "left"],
-                   [ds_top, "top"]
                    ])
 
 
[…]

[…]
