### 2025-01-09 b455943
Remove unused icon links from documentation configuration
