### 2025-01-09 fc539ff
Add reactions module and refactor reaction force calculations
