+++ src/phasefieldx/Element/Phase_Field/__init__.py
@@ -4,9 +4,3 @@
 
 from .Input import *
 from .solver import *
-
-# Optionally, you can specify which symbols to export when using 'from
-# phasefieldx import *'
-__all__ = [
-    'solver',
-]
