+++ src/phasefieldx/Element/Phase_Field_Fracture/__init__.py
@@ -6,5 +6,4 @@ from .solver import *
 from .Input import *
 from .fatigue_degradation_functions import *
 from .g_degradation_functions import *
-from .reactions_forces_functions import *
 from .split_energy_stress_tangent_functions import *
