+++ src/phasefieldx/Element/Elasticity/__init__.py
@@ -3,9 +3,4 @@
 # Import submodules to be included in the package namespace
 
 from .Input import *
-from .reactions_forces_functions import *
-
-__all__ = [
-    'Input',
-    'reactions_forces_functions'
-]
+from .solver import *
