+++ src/phasefieldx/Reactions/__init__.py
@@ -1,0 +1,3 @@
+# phasefieldx/Reactions/__init__.py
+
+from .reactions_forces import *

[…]
