### 2025-01-09 fd612d7
Bump version to 0.0.7
