### 2025-01-09 ac01354
Update Python version to 3.10 in configuration files and dependencies
