### 2025-01-09 bb12e6e
Refactor logging format in NewtonSolver for improved readability
