### 2025-01-09 e7869bd
Update test cases to correct facet marker labels for phase field simulations
