+++ test/Element/Phase_Field_Fracture/test_phase_field_fracture_element_anisotropic.py
@@ -64,8 +64,8 @@ def test_phase_field_simulation_anisotropic(split_energy_i):
     ds_top = get_ds_bound_from_marker(top_facet_marker, msh, fdim)
 
     ds_list = np.array([
+        [ds_top, "top"],
         [ds_bottom, "bottom"],
-        [ds_top, "top"]
     ])
 
     ###############################################################################
