### 2025-01-09 b044795
Remove 1D simulation from reaction forces test and update fixture return values
