+++ test/Reactions/test_reaction_forces.py
@@ -14,70 +14,6 @@ from phasefieldx.Boundary.boundary_conditions import bc_x, bc_xy, bc_xyz, get_ds
 from phasefieldx.PostProcessing.ReferenceResult import AllResults
 
 
-def run_1d_simulation():
-    Data = Input(E=210.0,
-                nu=0.3,
-                save_solution_xdmf=False,
-                save_solution_vtu=True,
-                results_folder_name="1_dimension")
-    divx = 1
-    lx  = 1
-    msh = dolfinx.mesh.create_unit_interval(mpi4py.MPI.COMM_WORLD,divx)
-
-
-    def bottom(x):
-        return np.isclose(x[0], 0)
-
-    def top(x):
-        return np.isclose(x[0], lx)
-
-    fdim = msh.topology.dim - 1 # Dimension of the mesh facets
-
-    bottom_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, bottom)
-    top_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, top)
-
-    ds_bottom = get_ds_bound_from_marker(bottom_facet_marker, msh, fdim)
-    ds_top = get_ds_bound_from_marker(top_facet_marker, msh, fdim)
-
-    ds_list = np.array([
-                    [ds_top, "top"],
-                    [ds_bottom, "bottom"]
-                    ])
-
-    V_u = dolfinx.fem.functionspace(msh, ("Lagrange", 1, (msh.geometry.dim, )))
-
-    bc_bottom = bc_x(bottom_facet_marker, V_u, fdim, value=0.0)
-    bc_top = bc_x(top_facet_marker, V_u, fdim, value=0.0)
-
-    bcs_list_u = [bc_top, bc_bottom]
-
-    def update_boundary_conditions(bcs, time):
-        val = 0.0003 * time
-        bcs[0].g.value[...] = petsc4py.PETSc.ScalarType(val)
-        return val, 0, 0
-
-    T_list_u = None
-    update_loading = None
-    f = None
-
-    final_time = 10.0
-    dt = 1.0
-
-    solve(Data,
-        msh,
-        final_time,
-        V_u,
-        bcs_list_u,
-        update_boundary_conditions,
-        f,
-        T_list_u,
-        update_loading,
-        ds_list,
-        dt,
-        path=None,
-        quadrature_degree=2)
-
-
 def run_2d_simulation():
     Data = Input(E=210.0,
                 nu=0.3,
@@ -216,26 +152,21 @@ def run_3d_simulation():
 @pytest.fixture(scope="module")
 def run_simulations():
     # Run the simulations for 1D, 2D, and 3D cases
-    run_1d_simulation()
     run_2d_simulation()
     run_3d_simulation()
     # Load results for 1D, 2D, and 3D cases
-    S_1d = AllResults("1_dimension")
     S_2d = AllResults("2_dimension")
     S_3d = AllResults("3_dimension")
 
-    return S_1d, S_2d, S_3d
+    return S_2d, S_3d
 
 def test_reaction_forces(run_simulations):
 
-    S_1d, S_2d, S_3d = run_simulations
+    S_2d, S_3d = run_simulations
 
     # Extract reaction forces
-    reaction_1d = S_1d.reaction_files['bottom.reaction']["Rx"]
     reaction_2d = S_2d.reaction_files['bottom.reaction']["Ry"]
     reaction_3d = S_3d.reaction_files['bottom.reaction']["Ry"]
 
     # Check if the reaction forces are equal
[…]
