### 2025-01-09 4fecfa1
Fix boundary condition application in reaction forces calculation
