+++ pyproject.toml
@@ -15,7 +15,7 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
-dependencies = ["fenics-dolfinx", "gmsh", "mpi4py", "numpy", "pandas", "pyvista"]
+
 keywords = ["phase field", "phase-field", "fracture", "fatigue", "solidification", "allen cahn", "finite element"]
 
 [tool.hatch.version]
