### 2025-01-09 ea72954
Specify exact Python version in environment configuration files
