+++ docs/environment.yml
@@ -2,7 +2,7 @@ name: phasefieldx-docs-env
 channels:
   - conda-forge
 dependencies:
-  - python>=3.10
+  - python=3.10
   - pyvista
   - jupyterlab
   - numpy
