### 2025-01-09 d3a352d
Update Miniconda setup action to version 3 and ensure Python version is quoted
