+++ .github/workflows/testing.yml
@@ -17,9 +17,9 @@ jobs:
       uses: actions/checkout@v3
 
     - name: Set up Conda
-      uses: conda-incubator/setup-miniconda@v2  
+      uses: conda-incubator/setup-miniconda@v3
       with:
-        python-version: 3.10
+        python-version: "3.10"
         auto-update-conda: true
         auto-activate-base: false
 
