### 2025-01-09 d61656a
Refactor logging in NewtonSolver to improve readability of convergence criterion
