### 2025-01-09 4f3fb5a
Fix formatting in NewtonSolver configuration string for improved readability
