### 2025-01-09 ae9d088
Improve logging format for better readability in Input and library_versions modules
