### 2025-01-09 a34e1a8
Fix formatting in error messages for improved readability in invariants and test files
