### 2025-01-09 4fb6279
Fix formatting in fatigue degradation function output for improved readability
