### 2025-01-09 5659076
Fix formatting in phase field test assertions for improved readability
