### 2025-01-09 587bc4d
Refactor assertions in phase field tests for improved accuracy and readability
