### 2025-01-09 95de853
Bump version to 0.0.8
