### 2025-01-09 CastillonMiguel/phasefieldx#9 @CastillonMiguel
Thanks for the feedback! I’ve resolved the plot display issue and improved text formatting in that file and across other related files for better readability
