### 2025-01-16 ad42f8b
Fix typo in external energy term notation in phase-field fracture model equation
