+++ joss/paper.md
@@ -31,7 +31,7 @@ The phase-field fracture model was extended by [@Miehe2010] to incorporate a reg
 
 A diffuse variable, the “phase-field” $\phi$, distinguishes between intact ($\phi = 0$) and fractured ($\phi = 1$) regions, with a smooth transition between them. This setup allows cracks to evolve based on energy minimization principles, eliminating the need for explicit crack-path tracking. The functional is given by \autoref{eq:phase_field_fracture_functional},
 \begin{equation}\label{eq:phase_field_fracture_functional}
-\mathcal{E}(\boldsymbol u, \phi) = \int_\Omega g(\phi)  \psi(\epsilon(\boldsymbol u)) \mathrm{d}\Omega + G_c \int_\Omega \left( \frac{1}{2l}\phi^2 + \frac{l}{2} |\nabla \phi|^2 \right) \mathrm{d}\Omega - E_{\text{ext}}[\boldsymbol u].
+\mathcal{E}(\boldsymbol u, \phi) = \int_\Omega g(\phi)  \psi(\epsilon(\boldsymbol u)) \mathrm{d}\Omega + G_c \int_\Omega \left( \frac{1}{2l}\phi^2 + \frac{l}{2} |\nabla \phi|^2 \right) \mathrm{d}\Omega - \mathcal{E}_{\text{ext}}[\boldsymbol u].
 \end{equation}
 where $\phi$ is the phase-field variable, $\boldsymbol u$ is the displacement vector field, $g(\phi) = (1 - \phi)^2$ is the degradation function that reduces material stiffness as damage progresses, $\epsilon(\boldsymbol u)= \frac{1}{2} \left( \nabla \boldsymbol u + \nabla^T \boldsymbol u \right)$ is the small strain tensor, $\psi(\epsilon(\boldsymbol u)) = \frac{1}{2}\lambda tr^2(\epsilon(\boldsymbol u)) + \mu tr(\epsilon(\boldsymbol u)^2)$ is the strain energy, $G_c$ is the critical energy release rate, $l$ is the length scale parameter that controls the width of the diffuse crack region and $E_{\text{ext}}[\boldsymbol u] = \int_\Omega f \cdot \boldsymbol u \, \mathrm{d}\Omega + \int_{\partial \Omega} t \cdot \boldsymbol u \, \mathrm{d}S$ represents the external terms, with $f$ being the prescribed volume force in $\Omega$ and $t$ the surface traction force on $\partial \Omega$.
 
