### 2025-01-24 d3dc00d
Add Gmsh geometry and mesh generation files for double edge notch tension test
