### 2025-01-24 647f302
Clarify notation in phase-field fracture model equation in paper documentation
