+++ joss/paper.md
@@ -33,7 +33,7 @@ A diffuse variable, the “phase-field” $\phi$, distinguishes between intact (
 \begin{equation}\label{eq:phase_field_fracture_functional}
 \mathcal{E}(\boldsymbol u, \phi) = \int_\Omega g(\phi)  \psi(\epsilon(\boldsymbol u)) \mathrm{d}\Omega + G_c \int_\Omega \left( \frac{1}{2l}\phi^2 + \frac{l}{2} |\nabla \phi|^2 \right) \mathrm{d}\Omega - \mathcal{E}_{\text{ext}}[\boldsymbol u].
 \end{equation}
-where $\phi$ is the phase-field variable, $\boldsymbol u$ is the displacement vector field, $g(\phi) = (1 - \phi)^2$ is the degradation function that reduces material stiffness as damage progresses, $\epsilon(\boldsymbol u)= \frac{1}{2} \left( \nabla \boldsymbol u + \nabla^T \boldsymbol u \right)$ is the small strain tensor, $\psi(\epsilon(\boldsymbol u)) = \frac{1}{2}\lambda tr^2(\epsilon(\boldsymbol u)) + \mu tr(\epsilon(\boldsymbol u)^2)$ is the strain energy, $G_c$ is the critical energy release rate, $l$ is the length scale parameter that controls the width of the diffuse crack region and $E_{\text{ext}}[\boldsymbol u] = \int_\Omega f \cdot \boldsymbol u \, \mathrm{d}\Omega + \int_{\partial \Omega} t \cdot \boldsymbol u \, \mathrm{d}S$ represents the external terms, with $f$ being the prescribed volume force in $\Omega$ and $t$ the surface traction force on $\partial \Omega$.
+where $\phi$ is the phase-field variable, $\boldsymbol u$ is the displacement vector field, $g(\phi) = (1 - \phi)^2$ is the degradation function that reduces material stiffness as damage progresses, $\epsilon(\boldsymbol u)= \frac{1}{2} \left( \nabla \boldsymbol u + \nabla^T \boldsymbol u \right)$ is the small strain tensor, $\psi(\epsilon(\boldsymbol u)) = \frac{1}{2}\lambda tr^2(\epsilon(\boldsymbol u)) + \mu tr(\epsilon(\boldsymbol u)^2)$ is the strain energy, $G_c$ is the critical energy release rate, $l$ is the length scale parameter that controls the width of the diffuse crack region and $\mathcal{E}_{\text{ext}}[\boldsymbol u] = \int_\Omega f \cdot \boldsymbol u \, \mathrm{d}\Omega + \int_{\partial \Omega} t \cdot \boldsymbol u \, \mathrm{d}S$ represents the external terms, with $f$ being the prescribed volume force in $\Omega$ and $t$ the surface traction force on $\partial \Omega$.
 
 The energy can be split into two components $\psi(\epsilon(\boldsymbol u)) = \psi_a(\epsilon(\boldsymbol u)) + \psi_b(\epsilon(\boldsymbol u))$, where $\psi_a$ represents the energy associated with tensile stresses capable of generating fractures, and the degradation function is applied only to this component. This yields the form $g(\phi) \psi_a(\epsilon(\boldsymbol u)) + \psi_b(\epsilon(\boldsymbol u))$, preventing crack formation due to compressive forces. Various methods exist for performing this energy split, such as the spectral decomposition [@Miehe2010] and the volumetric-deviatoric decomposition [@Amor2009]. Models that incorporate this energy split are known as anisotropic models, while models without energy splitting are referred to as isotropic models.
 
