### 2025-03-03 b66ab35
Add examples for tensile specimen and ASTM standard mesh generation with visualization
