### 2025-03-03 c0cb25d
Enhance elasticity solver by adding boundary condition names for improved clarity
