### 2025-03-03 c1d80c7
Add boundary condition names to phase-field fracture examples for clarity
