+++ examples/PhaseFieldFracture/plot_1700.py
@@ -204,6 +204,7 @@ bc_top = bc_xy(top_facet_marker, V_u, fdim)
 # field $\boldsymbol u$. This list facilitates easy management of multiple boundary
 # conditions and can be expanded if additional conditions are needed.
 bcs_list_u = [bc_top, bc_bottom]
+bcs_list_u_names = ["top", "bottom"]
 
 ###############################################################################
 # Function: `update_boundary_conditions`
@@ -317,7 +318,8 @@ solve(Data,
       update_loading,
       ds_list,
       dt,
-      path=None)
+      path=None,
+      bcs_list_u_names=bcs_list_u_names)
 
 
 ###############################################################################
