+++ examples/PhaseFieldFracture/plot_1714.py
@@ -215,6 +215,7 @@ bc_top = bc_y(top_facet_marker, V_u, fdim)
 # field $\boldsymbol u$. This list facilitates easy management of multiple boundary
 # conditions and can be expanded if additional conditions are needed.
 bcs_list_u = [bc_top, bc_bottom_left, bc_bottom_right]
+bcs_list_u_names = ["top", "bottom_left", "bottom_right"]
 
 ###############################################################################
 # Function: `update_boundary_conditions`
@@ -326,7 +327,8 @@ dt = 1
 #       update_loading,
 #       ds_list,
 #       dt,
-#       path=None)
+#       path=None,
+#       bcs_list_u_names=bcs_list_u_names)
 
 
 ###############################################################################
