+++ examples/PhaseFieldFracture/plot_1712.py
@@ -193,6 +193,7 @@ bc_right = bc_y(right_facet_marker, V_u, fdim)
 # field $\boldsymbol u$. This list facilitates easy management of multiple boundary
 # conditions and can be expanded if additional conditions are needed.
 bcs_list_u = [bc_top, bc_bottom, bc_left, bc_right]
+bcs_list_u_names = ["top", "bottom", "left", "right"]
 
 ###############################################################################
 # Function: `update_boundary_conditions`
@@ -306,7 +307,8 @@ final_time = 150.0
 #       update_loading,
 #       ds_list,
 #       dt,
-#       path=None)
+#       path=None,
+#       bcs_list_u_names=bcs_list_u_names)
 
 
 ###############################################################################
