+++ examples/PhaseFieldFracture/plot_1715.py
@@ -214,6 +214,7 @@ bc_top = bc_y(top_facet_marker, V_u, fdim)
 # field $\boldsymbol u$. This list facilitates easy management of multiple boundary
 # conditions and can be expanded if additional conditions are needed.
 bcs_list_u = [bc_top, bc_bottom, bc_left]
+bcs_list_u_names = ["top", "bottom", "left"]
 
 ###############################################################################
 # Function: `update_boundary_conditions`
@@ -327,7 +328,8 @@ final_time = 200.0
 #       update_loading,
 #       ds_list,
 #       dt,
-#       path=None)
+#       path=None,
+#       bcs_list_u_names=bcs_list_u_names)
 
 
 ###############################################################################
