### 2025-03-03 25b6644
Add boundary condition names to fatigue plot for clarity
