### 2025-03-03 b133f20
Add boundary condition names to phase-field fracture tests for clarity
