+++ test/Element/Phase_Field_Fracture/test_phase_field_fracture_element.py
@@ -79,7 +79,8 @@ def test_phase_field_simulation():
     bc_bottom = bc_xy(bottom_facet_marker, V_u, fdim)
     bc_top = bc_xy(top_facet_marker, V_u, fdim)
     bcs_list_u = [bc_top, bc_bottom]
-
+    bcs_list_u_names = ["top", "bottom"]
+    
     def update_boundary_conditions(bcs, time):
         if time <= 50:
             val = 0.0003 * time
@@ -117,7 +118,8 @@ def test_phase_field_simulation():
           update_loading,
           ds_list,
           dt,
-          path=None)
+          path=None,
+          bcs_list_u_names=bcs_list_u_names)
 
     S = AllResults(Data.results_folder_name)
     displacement = S.dof_files["top.dof"]["Uy"]
@@ -127,7 +129,7 @@ def test_phase_field_simulation():
 
     np.testing.assert_allclose(S.reaction_files["top.reaction"]["Ry"], sigma_t, rtol=1e-3, atol=1e-8, 
                            err_msg=f"Computed reaction force {S.reaction_files['top.reaction']['Ry']} does not match theoretical reaction {sigma_t}")
-    
+   
     # Clean up: Remove generated files
     if os.path.exists(Data.results_folder_name):
         import shutil
