### 2025-03-03 7f397df
Refactor stagger iteration parameters in phase-field fracture solver for improved clarity and flexibility
