### 2025-03-03 a76170e
Restore stagger iteration parameters in phase-field fracture tests for consistency
