+++ test/Element/Phase_Field_Fracture/test_phase_field_fracture_element_anisotropic.py
@@ -31,9 +31,6 @@ def test_phase_field_simulation_anisotropic(split_energy_i):
                  fatigue_degradation_function="asymptotic",
                  fatigue_val=0.05625,
                  k=0.0,
-                 min_stagger_iter=2,
-                 max_stagger_iter=500,
-                 stagger_error_tol=1e-8,
                  save_solution_xdmf=False,
                  save_solution_vtu=True,
                  results_folder_name="1700_One_element_isotropic_tension_test_anisotropic")
@@ -123,7 +120,10 @@ def test_phase_field_simulation_anisotropic(split_energy_i):
           ds_list,
           dt,
           path=None,
-          bcs_list_u_names=bcs_list_u_names)
+          bcs_list_u_names=bcs_list_u_names,
+          min_stagger_iter=2,
+          max_stagger_iter=500,
+          stagger_error_tol=1e-8)
 
     S = AllResults(Data.results_folder_name)
     displacement = S.dof_files["top.dof"]["Uy"]
