+++ test/Element/Phase_Field_Fracture/test_phase_field_fracture_element.py
@@ -29,9 +29,6 @@ def test_phase_field_simulation():
                  fatigue_degradation_function="asymptotic",
                  fatigue_val=0.05625,
                  k=0.0,
-                 min_stagger_iter=2,
-                 max_stagger_iter=500,
-                 stagger_error_tol=1e-8,
                  save_solution_xdmf=False,
                  save_solution_vtu=True,
                  results_folder_name="1700_One_element_isotropic_tension_test")
@@ -119,7 +116,10 @@ def test_phase_field_simulation():
           ds_list,
           dt,
           path=None,
-          bcs_list_u_names=bcs_list_u_names)
+          bcs_list_u_names=bcs_list_u_names,
+          min_stagger_iter=2,
+          max_stagger_iter=500,
+          stagger_error_tol=1e-8)
 
     S = AllResults(Data.results_folder_name)
     displacement = S.dof_files["top.dof"]["Uy"]
