### 2025-03-03 33db5b5
Remove stagger iteration parameters from fatigue and phase-field fracture examples for consistency
