+++ examples/PhaseFieldFracture/plot_1714.py
@@ -103,9 +103,6 @@ Data = Input(E=20.8,     # young modulus
              fatigue_degradation_function="asymptotic",
              fatigue_val=0.05625,
              k=0.0,
-             min_stagger_iter=2,
-             max_stagger_iter=1000,
-             stagger_error_tol=1e-6,
              save_solution_xdmf=False,
              save_solution_vtu=True,
              results_folder_name="1714_Three_point_bending")
@@ -328,7 +325,10 @@ dt = 1
 #       ds_list,
 #       dt,
 #       path=None,
-#       bcs_list_u_names=bcs_list_u_names)
+#       bcs_list_u_names=bcs_list_u_names,
+#       min_stagger_iter=2,
+#       max_stagger_iter=1000,
+#       stagger_error_tol=1e-6)
 
 
 ###############################################################################
