+++ examples/PhaseFieldFracture/plot_1702.py
@@ -111,9 +111,6 @@ Data = Input(E=210.0,   # young modulus
              fatigue_degradation_function="asymptotic",
              fatigue_val=0.0,
              k=0.0,
-             min_stagger_iter=2,
-             max_stagger_iter=500,
-             stagger_error_tol=1e-8,
              save_solution_xdmf=False,
              save_solution_vtu=True,
              results_folder_name="1702_One_element_anisotropic_spectral")
@@ -317,7 +314,10 @@ solve(Data,
       ds_list,
       dt,
       path=None,
-      bcs_list_u_names=bcs_list_u_names)
+      bcs_list_u_names=bcs_list_u_names,
+      min_stagger_iter=2,
+      max_stagger_iter=500,
+      stagger_error_tol=1e-8,)
 
 
 ###############################################################################
