+++ examples/PhaseFieldFracture/plot_1715.py
@@ -102,9 +102,6 @@ Data = Input(E=210.0,   # young modulus
              fatigue_degradation_function="asymptotic",
              fatigue_val=0.0,
              k=0.0,
-             min_stagger_iter=2,
-             max_stagger_iter=600,
-             stagger_error_tol=1e-8,
              save_solution_xdmf=False,
              save_solution_vtu=True,
              results_folder_name="1715_Symmetry_Center_notched_tension_test")
@@ -329,7 +326,10 @@ final_time = 200.0
 #       ds_list,
 #       dt,
 #       path=None,
-#       bcs_list_u_names=bcs_list_u_names)
+#       bcs_list_u_names=bcs_list_u_names,
+#       min_stagger_iter=2,
+#       max_stagger_iter=600,
+#       stagger_error_tol=1e-8)
 
 
 ###############################################################################
