+++ examples/Fatigue/plot_1800.py
@@ -111,9 +111,6 @@ Data = Input(E=210.0,    # young modulus
              fatigue_degradation_function="asymptotic",
              fatigue_val=0.05625,
              k=0.0,
-             min_stagger_iter=2,
-             max_stagger_iter=500,
-             stagger_error_tol=1e-8,
              save_solution_xdmf=False,
              save_solution_vtu=True,
              results_folder_name="1800_Fatigue_Single_Edge_Notched_Tension_Test")
@@ -340,7 +337,10 @@ final_time = 8 * 200 + 1
 #       ds_list,
 #       dt,
 #       path=None,
-#       bcs_list_u_names=bcs_list_u_names)
+#       bcs_list_u_names=bcs_list_u_names,
+#       min_stagger_iter=2,
+#       max_stagger_iter=500,
+#       stagger_error_tol=1e-8)
 
 
 ###############################################################################
