### 2025-03-03 #7307 @IgorBaratta
Hi Bonan Zhu,
I sincerely apologize for the delayed response. This somehow slipped off my
radar, and I just realized that the email had been going to my spam folder.
That said, I'm happy to recommend it for publication. PhaseFieldX is
well-documented, easy to install, and straightforward to use.
