+++ test/Element/Phase_Field_Fracture/test_phase_field_fracture_element.py
@@ -77,7 +77,7 @@ def test_phase_field_simulation():
     bc_top = bc_xy(top_facet_marker, V_u, fdim)
     bcs_list_u = [bc_top, bc_bottom]
     bcs_list_u_names = ["top", "bottom"]
-    
+
     def update_boundary_conditions(bcs, time):
         if time <= 50:
             val = 0.0003 * time
@@ -127,9 +127,9 @@ def test_phase_field_simulation():
     phi_t = 2 * psi_t / (Data.Gc / Data.l + 2 * psi_t)
     sigma_t = displacement * (Data.lambda_ + 2 * Data.mu) * (1 - phi_t)**2
 
-    np.testing.assert_allclose(S.reaction_files["top.reaction"]["Ry"], sigma_t, rtol=1e-3, atol=1e-8, 
-                           err_msg=f"Computed reaction force {S.reaction_files['top.reaction']['Ry']} does not match theoretical reaction {sigma_t}")
-   
+    np.testing.assert_allclose(S.reaction_files["top.reaction"]["Ry"], sigma_t, rtol=1e-3, atol=1e-8,
+                               err_msg=f"Computed reaction force {S.reaction_files['top.reaction']['Ry']} does not match theoretical reaction {sigma_t}")
+
     # Clean up: Remove generated files
     if os.path.exists(Data.results_folder_name):
         import shutil
