+++ test/Reactions/test_reaction_forces.py
@@ -16,16 +16,16 @@ from phasefieldx.PostProcessing.ReferenceResult import AllResults
 
 def run_2d_simulation():
     Data = Input(E=210.0,
-                nu=0.3,
-                save_solution_xdmf=False,
-                save_solution_vtu=True,
-                results_folder_name="2_dimension")
+                 nu=0.3,
+                 save_solution_xdmf=False,
+                 save_solution_vtu=True,
+                 results_folder_name="2_dimension")
 
     divx, divy = 1, 1
     lx, ly = 1.0, 1.0
     msh = dolfinx.mesh.create_rectangle(mpi4py.MPI.COMM_WORLD,
                                         [np.array([0, 0]),
-                                        np.array([lx, ly])],
+                                         np.array([lx, ly])],
                                         [divx, divy],
                                         cell_type=dolfinx.mesh.CellType.quadrilateral)
 
@@ -35,27 +35,27 @@ def run_2d_simulation():
     def top(x):
         return np.isclose(x[1], ly)
 
-    fdim = msh.topology.dim - 1 
-    
+    fdim = msh.topology.dim - 1
+
     bottom_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, bottom)
     top_facet_marker = dolfinx.mesh.locate_entities_boundary(msh, fdim, top)
-    
+
     ds_bottom = get_ds_bound_from_marker(bottom_facet_marker, msh, fdim)
     ds_top = get_ds_bound_from_marker(top_facet_marker, msh, fdim)
-    
+
     ds_list = np.array([
-                    [ds_top, "top"],
-                    [ds_bottom, "bottom"],
-                    ])
+        [ds_top, "top"],
+        [ds_bottom, "bottom"],
+    ])
 
     V_u = dolfinx.fem.functionspace(msh, ("Lagrange", 1, (msh.geometry.dim, )))
 
     bc_bottom = bc_xy(bottom_facet_marker, V_u, fdim, value_x=0.0, value_y=0.0)
     bc_top = bc_xy(top_facet_marker, V_u, fdim, value_x=0.0, value_y=0.0)
-    
+
     bcs_list_u = [bc_top, bc_bottom]
-    bcs_list_u_names= ["top", "bottom"]
-    
+    bcs_list_u_names = ["top", "bottom"]
+
     def update_boundary_conditions(bcs, time):
         val = 0.0003 * time
         bcs[0].g.value[1] = petsc4py.PETSc.ScalarType(val)
@@ -69,35 +69,35 @@ def run_2d_simulation():
     dt = 1.0
 
     solve(Data,
-        msh,
-        final_time,
-        V_u,
-        bcs_list_u,
-        update_boundary_conditions,
-        f,
-        T_list_u,
-        update_loading,
-        ds_list,
-        dt,
-        path=None,
-        quadrature_degree=2,
-        bcs_list_u_names=bcs_list_u_names)
+          msh,
+          final_time,
+          V_u,
+          bcs_list_u,
+          update_boundary_conditions,
+          f,
+          T_list_u,
+          update_loading,
+          ds_list,
+          dt,
+          path=None,
+          quadrature_degree=2,
+          bcs_list_u_names=bcs_list_u_names)
 
 
 def run_3d_simulation():
     Data = Input(E=210.0,
-                nu=0.3,
-                save_solution_xdmf=False,
-                save_solution_vtu=True,
-                results_folder_name="3_dimension")
+                 nu=0.3,
+                 save_solution_xdmf=False,
+                 save_solution_vtu=True,
+                 results_folder_name="3_dimension")
 
[…]
