### 2025-03-24 #7307 @zhubonan
Hi  @CastillonMiguel, I am happy to have this work accepted by JOSS but before we proceed to the next stages there are a few things for you to complete:

The main thing is to have the released so the reviewed version of the software will have a DOI. Please give me a ping once it is done. Thanks!
