### 2025-03-24 ebbc306
Corrected G_c unit descriptions in multiple examples, changing from kN/mm² to kN/mm.
