+++ examples/PhaseFieldFracture/plot_1714.py
@@ -35,7 +35,7 @@ The Young's modulus, Poisson's ratio, and the critical energy release rate are g
 +----+---------+--------+
 | nu | 0.3     | [-]    |
 +----+---------+--------+
-| Gc | 0.0005  | kN/mm2 |
+| Gc | 0.0005  | kN/mm  |
 +----+---------+--------+
 | l  | 0.006   | mm     |
 +----+---------+--------+
