+++ examples/PhaseFieldFracture/plot_1702.py
@@ -36,7 +36,7 @@ The Young's modulus, Poisson's ratio, and the critical energy release rate are g
    +----+---------+--------+
    | nu | 0.3     | [-]    |
    +----+---------+--------+
-   | Gc | 0.005   | kN/mm2 |
+   | Gc | 0.005   | kN/mm  |
    +----+---------+--------+
    | l  | 0.1     | mm     |
    +----+---------+--------+
