### 2025-03-24 c1cfa56
Bump version to 0.1.2 for release
