### 2025-03-24 #7307 @CastillonMiguel
Hi @zhubonan ,

I have completed the requested steps:

- [x] Released the software with the latest changes: [v0.1.2](https://github.com/CastillonMiguel/phasefieldx/tree/v0.1.2)
- [x] Archived the release on Zenodo → [DOI: 10.5281/zenodo.15077939](https://doi.org/10.5281/zenodo.15077939)
- [x] Confirmed that the title and author list (including ORCIDs) match the JOSS paper
- [x] Verified that the license in the archive matches the software license (MIT license)

Everything should now be in place for the next steps. Please let me know if any additional adjustments are needed. Thanks again!
