### 2025-04-08 #7307 @kyleniemeyer
Hello @CastillonMiguel, I made some light fixes to reference formatting in the text where the author name should be part of the sentence. Could you review and merge? https://github.com/CastillonMiguel/phasefieldx/pull/15
