### 2026-01-27 #7307 @CastillonMiguel
Thank you for your message and your interest in PhaseFieldX. Currently, there is no video tutorial available, but I plan to expand the documentation and provide more detailed guidance on using the software over the coming months.

Please note that this issue is related to the JOSS publication itself, so practical questions about using the software are best addressed directly on the PhaseFieldX GitHub repository:
https://github.com/CastillonMiguel/phasefieldx
